% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-utils.R
\name{cfg_beta_prob_rng}
\alias{cfg_beta_prob_rng}
\title{Generate a random probability based on features of the simulation}
\usage{
cfg_beta_prob_rng(probability_fn = ~0.8, kappa_fn = ~0.1)
}
\arguments{
\item{probability_fn}{a function which gives the time-varying mean of a beta
distribution, The function will be called minimally with \code{.x}\`t`` which
will be the time as a time period. Other variables may be present.}

\item{kappa_fn}{a function which gives the time-varying dispersion of a beta
distribution. The function will be called with minimally \code{.x}\\code{t} which
will be the time period and \code{.y}\\code{mean} will be the mean. Other variables
may be present.}
}
\value{
a time dependent function that inputs a time (as a time_period) and
returns an probability for each day defined by the \code{probability_fn} and \code{kappa_fn}
}
\description{
Generate a random probability based on features of the simulation
}
\examples{
fn = cfg_beta_prob_rng(~ ifelse(.x<=5,0.1,0.9))
fn(1:10)
}
\seealso{
\code{\link[=rbeta2]{rbeta2()}}
}
