% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-utils.R
\name{cfg_ip_sampler_rng}
\alias{cfg_ip_sampler_rng}
\title{Randomly sample from an empirical distribution}
\usage{
cfg_ip_sampler_rng(ip = i_empirical_ip)
}
\arguments{
\item{ip}{a long format empirical distribution

A dataframe containing the following columns: 
\itemize{
\item boot (anything + default(1)) - a bootstrap identifier
\item probability (proportion) - the probability of new event during this period.
\item a0 (double) - the beginning of the time period (in days)
\item a1 (double) - the end of the time period (in days)
}

Must be grouped by: boot (exactly).

A default value is defined.}
}
\value{
a function which accepts \code{n} parameter which produces random samples
from the \code{ip} distribution
}
\description{
This is used for random sampling from the infectivity profile for times to
infection, for example. There is nothing to stop you putting in a delay
distribution with negative times but strange things may happen in your
simulation.
}
\examples{
tmp = cfg_ip_sampler_rng(ganyani_ip_2)(10000)

# This discretised ganyani distribution is based on these figures:
# mean: 5.2 (3.78-6.78) and sd: 1.72 (0.91-3.93)
format_ip(ganyani_ip_2)

mean(tmp) # Should be about 5.2
stats::sd(tmp) # Should be about 1.72
}
