% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-utils.R
\name{cfg_linear_fn}
\alias{cfg_linear_fn}
\title{Linear function from dataframe}
\usage{
cfg_linear_fn(changes, ..., col_name = setdiff(colnames(changes), "t"))
}
\arguments{
\item{changes}{a dataframe with \code{t} and \verb{<col_name>} columns which define the change
points for a piecewise linear function.}

\item{col_name}{the value column (optional if only 2 columns)}
}
\value{
a function that inputs a vector \code{t} and returns a linearly
interpolated value from \verb{<col_name>}
}
\description{
Linear function from dataframe
}
\concept{test}
