% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-utils.R
\name{cfg_step_fn}
\alias{cfg_step_fn}
\title{Step function from dataframe}
\usage{
cfg_step_fn(changes, ..., col_name = setdiff(colnames(changes), "t"))
}
\arguments{
\item{changes}{a dataframe with \code{t} and \verb{<col_name>} columns which define the cut points
for a step function.}

\item{col_name}{the value column (optional if only 2 columns)}
}
\value{
a function that inputs a vector \code{t} and returns the next smallest
corresponding value in \verb{<col_name>} (or the first one)
}
\description{
Step function from dataframe
}
\concept{test}
