% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-utils.R
\name{cfg_transition_fn}
\alias{cfg_transition_fn}
\title{Sample from a multinomial transition matrix}
\usage{
cfg_transition_fn(transition)
}
\arguments{
\item{transition}{a transition matrix or long format dataframe. If a matrix
the columns should add to 1, and the column names are the input class. row names are
the output class. the data frame format must have input, output, and probability
columns.}
}
\value{
a function that given an input will return samples of the output
class according to the probability distributions.
}
\description{
This is particularly designed for use within the \code{fn_list_next_gen} parameter
of \code{\link[=sim_branching_process]{sim_branching_process()}} to allow age group contract matrices to be applied
for example (assuming a categorical age).
}
\examples{
age = rep(c("child","adult","elderly"),100)

fn = cfg_transition_fn(tibble::tribble(
  ~input, ~output, ~probability,
  "child", "child", 0.5,
  "child", "adult", 0.5,
  "adult", "child", 0.5,
  "adult", "adult", 0.3,
  "adult", "elderly", 0.2,
  "elderly","elderly", 0.5,
  "elderly","adult", 0.5,
))

table(fn(age),age)
}
\concept{test}
