% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-utils.R
\name{cfg_weekly_gamma_rng}
\alias{cfg_weekly_gamma_rng}
\title{Weekly delay function with day of week effect}
\usage{
cfg_weekly_gamma_rng(
  mean = c(1, 1, 1, 1, 4, 3, 2),
  sd = sqrt(mean),
  week_starts = weekdays(as.Date("2024-10-14"))
)
}
\arguments{
\item{mean}{a mean amount of delay for each day of the week}

\item{sd}{the SD of the delay}

\item{week_starts}{locale description of first day of week (default is a "Monday").}
}
\value{
a random number generator taking \code{t} time parameter and returning a
duration the that time \code{t} depending on the day of the week.
}
\description{
This function returns a random number generator from a gamma distribution
that has a weekly period and configurable degree of additional variability
around weekly pattern. It is suited to delays of things like testing which
may depend on the day of week.
}
\examples{
fn = cfg_weekly_gamma_rng(c(1,1,1,1,4,3,2))
matrix(fn(1:42),ncol=7,byrow=TRUE)
}
\concept{test}
