% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-utils.R
\name{cfg_weekly_ip_fn}
\alias{cfg_weekly_ip_fn}
\title{Weekly convolution distribution function}
\usage{
cfg_weekly_ip_fn(
  mean = c(1, 1, 1, 1, 4, 3, 2),
  sd = sqrt(mean),
  week_starts = weekdays(as.Date("2024-10-14"))
)
}
\arguments{
\item{mean}{The means of a gamma distributed delay function for each weekday}

\item{sd}{The sds of a gamma distributed delay function for each weekday}

\item{week_starts}{locale description of first day of week (default is a "Monday").}
}
\value{
a time dependent function that inputs a time (as a time_period) and
generates an IP delay distribution for each day varying by day of week
}
\description{
Weekly convolution distribution function
}
\examples{
cat(sapply(cfg_weekly_ip_fn()(1:7),format_ip),sep = "\n")
}
