% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-utils.R
\name{cfg_weekly_proportion_rng}
\alias{cfg_weekly_proportion_rng}
\title{Random probability function with day of week effect}
\usage{
cfg_weekly_proportion_rng(
  prob = c(0.8, 0.8, 0.8, 0.8, 0.8, 0.5, 0.5),
  kappa = 0.1,
  week_starts = weekdays(as.Date("2024-10-14"))
)
}
\arguments{
\item{prob}{the rates of e.g. ascertainment for each day of the week.}

\item{kappa}{dispersion parameter betwen 0 and 1. O is no dispersion. 1 is maximum}

\item{week_starts}{locale description of first day of week (default is a "Monday").}
}
\value{
a random number generator function taking \code{t} time parameter and
returning a probability of ascertainment for that time, depending on day of
week etc.
}
\description{
This function returns a random probability generator that has a weekly period
and configurable degree of additional variability around weekly pattern. It is
suited to probabilities of things like testing which may depend on the day
of week.
}
\examples{
fn = cfg_weekly_proportion_rng(c(0.9,0.9,0.9,0.9,0.9,0.1,0.1))
matrix(fn(1:42),ncol=7,byrow=TRUE)
}
\concept{test}
