% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_infectivity_profile}
\alias{covid_infectivity_profile}
\title{The covid_infectivity_profile dataset}
\format{
A dataframe containing the following columns:
\itemize{
\item tau (numeric) - the time column
\item boot (numeric) - the boot column
\item probability (numeric) - the probability column
}

Minimally grouped by: boot (and other groupings allowed).

1300 rows and 3 columns
}
\usage{
data(covid_infectivity_profile)
}
\description{
The covid_infectivity_profile dataset
}
\keyword{datasets}
