% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_ip}
\alias{covid_ip}
\title{A COVID-19 infectivity profile based on an empirical resampling approach}
\format{
A dataframe containing the following columns:
\itemize{
\item boot (anything + default(1)) - a bootstrap identifier
\item probability (proportion) - the probability of infection between previous time period until \code{time}
\item time (double) - the end of the time period (in days)
}

Must be grouped by: boot (exactly).

A dataframe containing the following columns:
\itemize{
\item boot (anything + default(1)) - a bootstrap identifier
\item probability (proportion) - the probability of infection between previous time period until \code{time}
\item tau (numeric) - the time index this probability relates to (in days)
\item a0 (numeric) - the beginning of the time period
\item a1 (numeric) - the end of the time period
}

Grouped by: boot.

1400 rows and 5 columns
}
\usage{
data(covid_ip)
}
\description{
An infectivity profile derived from a meta-analysis of serial intervals.
}
\concept{datasets}
