% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_test_sensitivity}
\alias{covid_test_sensitivity}
\title{Test sensitivity of PCR tests}
\format{
A dataframe containing the following columns:
\itemize{
\item tau (numeric) - the time column
\item probability (numeric) - the probability column
\item boot (integer) - the boot column
}

Must be grouped by: boot (and other groupings allowed).

5100 rows and 3 columns
}
\usage{
data(covid_test_sensitivity)
}
\description{
The probability of detecting COVID using PCR given time since infection, based on
Binny et al 2023.
}
\references{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC9384503/}
}
\concept{datasets}
