% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_viral_shedding}
\alias{covid_viral_shedding}
\title{The COVID-19 viral shedding duration}
\format{
A dataframe containing the following columns:
\itemize{
\item boot (anything + default(1)) - a bootstrap identifier
\item probability (proportion) - the probability of infection between previous time period until \code{time}
\item tau (numeric) - the time index this probability relates to (in days)
\item a0 (numeric) - the beginning of the time period
\item a1 (numeric) - the end of the time period
}

Grouped by: boot.

2600 rows and 3 columns
}
\usage{
data(covid_viral_shedding)
}
\description{
The COVID-19 viral shedding duration
}
\references{
\url{https://www.nature.com/articles/s41467-020-20568-4}
From Von Kampen et al.
}
\concept{datasets}
