% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic-times.R
\name{.day_interval}
\alias{.day_interval}
\title{Guess the intervals between a sequence of dates:}
\usage{
.day_interval(dates)
}
\arguments{
\item{dates}{a set of dates}
}
\value{
a natural number of days between ordering
}
\description{
Guess the intervals between a sequence of dates:
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

r = c(1,rpois(10, 3)+1)
interval = 2
dates = as.Date("2025-01-01")+r*interval
testthat::expect_equal(.day_interval(dates) == interval, TRUE)

interval = 0.5
dates = as.Date("2025-01-01")+r*interval
testthat::expect_equal(.day_interval(dates) == interval, TRUE)

}\if{html}{\out{</div>}}
}

