% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cdf.R
\name{.fit_lm_1d}
\alias{.fit_lm_1d}
\title{Fit a weighted 1D linear model and predict output}
\usage{
.fit_lm_1d(y, x)
}
\arguments{
\item{y}{the y values. At least 2.}

\item{x}{the x values. At least 2.}
}
\value{
a vector with intercept (\code{c}) and gradient (\code{m})
}
\description{
If all the weights are NA they are ignored.
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

testthat::expect_equal(.fit_lm_1d(y = 1:10 * 2 + 5, x = 1:10), c(c = 5, m = 2))
}\if{html}{\out{</div>}}
}

