% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-data.R
\name{.infer_grouping}
\alias{.infer_grouping}
\title{Add grouping to ensure a column is groupwise unique}
\usage{
.infer_grouping(df, unique_col)
}
\arguments{
\item{df}{a dataframe, maybe grouped}

\item{unique_col}{a column reference that should be groupwise unique}
}
\value{
a regrouped dataframe in which \code{unique_col} is indeed unique
}
\description{
Add grouping to ensure a column is groupwise unique
}
\keyword{internal}

\section{Unit tests}{
\if{html}{\out{<div class="sourceCode">}}\preformatted{

df = dplyr::tibble(
  original_grp = as.vector(sapply(LETTERS[1:4],rep,20)),
  to_find = as.vector(sapply(letters[1:10],rep,8)),
  to_make_unique = rep(1:10,8),
  distractor = runif(80)
) %>% dplyr::group_by(original_grp)
unique_col = as.symbol("to_make_unique")
tmp = .infer_grouping(df, to_make_unique)
testthat::expect_equal(
  format(groups(tmp)),
  c("original_grp", "to_find")
)
}\if{html}{\out{</div>}}
}

