% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-standalone-mixture-distributions.R
\name{.qmixgamma}
\alias{.qmixgamma}
\title{A quantile function for a mixture of gamma distributions}
\usage{
.qmixgamma(
  p,
  shapes,
  rates,
  weights = 1,
  na.rm = FALSE,
  method = c("exact", "samples", "moments"),
  ...
)
}
\arguments{
\item{p}{vector of probabilities.}

\item{shapes}{a vector of gamma distribution shapes}

\item{rates}{a vector of gamma distribution rates}

\item{weights}{a vector of weights}

\item{na.rm}{remove distributions with NA values for mean or sd}

\item{method}{one of \code{exact} (solve with uniroot), \code{samples} (random resampling), \code{moments} (Cornish Fisher approximation)}

\item{...}{passed on to internal function, \code{seed=XXX} to fix random seed}
}
\value{
the value of the \code{p}th quantile
}
\description{
A quantile function for a mixture of gamma distributions
}
\examples{
try({
  .qmixgamma(p=c(0.025,0.5,0.975), shapes=c(10,13,14), rates=c(1,1,2), method="moments")
  .qmixgamma(p=c(0.025,0.5,0.975), shapes=c(10,13,14), rates=c(1,1,2), method="exact")

  means = runif(100,5,6)
  sds = runif(100,2,3)
  shapes = means^2/sds^2
  rates = means/sds^2

  if (sd(means) < mean(sds)) message("mixture moments should be close")

  qgamma(c(0.025,0.5,0.975), shape =mean(means^2)/mean(sds^2), rate = mean(means)/mean(sds^2))
  system.time(
  .qmixgamma(p=c(0.025,0.5,0.975), shapes=shapes, rates=rates, method="moments")
  )
  system.time(
  .qmixgamma(p=c(0.025,0.5,0.975), shapes=shapes, rates=rates, method="exact")
  )

  means = runif(100,2,12)
  sds = runif(100,0.5,1)
  shapes = means^2/sds^2
  rates = means/sds^2

  if (sd(means) < mean(sds)) message("mixture moments should be close")

  qgamma(c(0.025,0.5,0.975), shape =mean(means^2)/mean(sds^2), rate = mean(means)/mean(sds^2))
  system.time(
  .qmixgamma(p=c(0.025,0.5,0.975), shapes=shapes, rates=rates, method="moments")
  )
  system.time(
  .qmixgamma(p=c(0.025,0.5,0.975), shapes=shapes, rates=rates, method="exact")
  )
  quantile(means, prob=c(0.025,0.5,0.975))

})
}
\concept{distributions}
\keyword{internal}
