% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-utils.R
\name{.ts_evaluate}
\alias{.ts_evaluate}
\title{Evaluate a function in a timeseries dataframe}
\usage{
.ts_evaluate(fn, df)
}
\arguments{
\item{fn}{a function.}

\item{df}{a dataframe with a numeric time column, plus other columns}
}
\value{
a vector, the result of applying the function to df.
}
\description{
N.B. only exported as used in one vignette for demo purposes
}
\examples{
test = dplyr::tibble(
  time = 1:10,
  class = rep(c("one","two"),5)
)
.ts_evaluate(\(t,class) {print(t); print(class); 1}, test )

}
\keyword{internal}
