% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_covid_poisson}
\alias{england_covid_poisson}
\title{The England COVID-19 poisson model dataset}
\format{
A dataframe containing the following columns:
\itemize{
\item time (as.time_period) - the time column
\item incidence.fit (numeric) - the incidence.fit column
\item incidence.se.fit (numeric) - the incidence.se.fit column
\item incidence.0.025 (numeric) - the incidence.0.025 column
\item incidence.0.05 (numeric) - the incidence.0.05 column
\item incidence.0.25 (numeric) - the incidence.0.25 column
\item incidence.0.5 (numeric) - the incidence.0.5 column
\item incidence.0.75 (numeric) - the incidence.0.75 column
\item incidence.0.95 (numeric) - the incidence.0.95 column
\item incidence.0.975 (numeric) - the incidence.0.975 column
\item growth.fit (numeric) - the growth.fit column
\item growth.se.fit (numeric) - the growth.se.fit column
\item growth.0.025 (numeric) - the growth.0.025 column
\item growth.0.05 (numeric) - the growth.0.05 column
\item growth.0.25 (numeric) - the growth.0.25 column
\item growth.0.5 (numeric) - the growth.0.5 column
\item growth.0.75 (numeric) - the growth.0.75 column
\item growth.0.95 (numeric) - the growth.0.95 column
\item growth.0.975 (numeric) - the growth.0.975 column
}

Any grouping allowed.

1410 rows and 19 columns
}
\usage{
data(england_covid_poisson)
}
\description{
This is the output of the following estimator, and is here to speed up some
examples:
}
\details{
\if{html}{\out{<div class="sourceCode R">}}\preformatted{england_covid \%>\% time_aggregate() \%>\% ggoutbreak::poisson_locfit_model(window=14)
}\if{html}{\out{</div>}}
}
\concept{datasets}
\keyword{datasets}
