% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_covid_poisson_age_stratified}
\alias{england_covid_poisson_age_stratified}
\title{The England COVID-19 age stratified poisson model dataset}
\format{
A dataframe containing the following columns:
\itemize{
\item class (enum(\verb{00_04},\verb{05_09},\verb{10_14},\verb{15_19},\verb{20_24},\verb{25_29},\verb{30_34},\verb{35_39},\verb{40_44},\verb{45_49},\verb{50_54},\verb{55_59},\verb{60_64},\verb{65_69},\verb{70_74},\verb{75_79},\verb{80_84},\verb{85_89},\verb{90+})) - the class column
\item time (as.time_period) - the time column
\item incidence.fit (numeric) - the incidence.fit column
\item incidence.se.fit (numeric) - the incidence.se.fit column
\item incidence.0.025 (numeric) - the incidence.0.025 column
\item incidence.0.05 (numeric) - the incidence.0.05 column
\item incidence.0.25 (numeric) - the incidence.0.25 column
\item incidence.0.5 (numeric) - the incidence.0.5 column
\item incidence.0.75 (numeric) - the incidence.0.75 column
\item incidence.0.95 (numeric) - the incidence.0.95 column
\item incidence.0.975 (numeric) - the incidence.0.975 column
\item growth.fit (numeric) - the growth.fit column
\item growth.se.fit (numeric) - the growth.se.fit column
\item growth.0.025 (numeric) - the growth.0.025 column
\item growth.0.05 (numeric) - the growth.0.05 column
\item growth.0.25 (numeric) - the growth.0.25 column
\item growth.0.5 (numeric) - the growth.0.5 column
\item growth.0.75 (numeric) - the growth.0.75 column
\item growth.0.95 (numeric) - the growth.0.95 column
\item growth.0.975 (numeric) - the growth.0.975 column
}

Minimally grouped by: class (and other groupings allowed).

26790 rows and 20 columns
}
\usage{
data(england_covid_poisson_age_stratified)
}
\description{
This is the output of the following estimator, and is here to speed up some
examples:
}
\details{
\if{html}{\out{<div class="sourceCode R">}}\preformatted{england_covid \%>\% ggoutbreak::poisson_locfit_model(window=14)
}\if{html}{\out{</div>}}
}
\concept{datasets}
\keyword{datasets}
