% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zz_pins.R
\name{england_covid_proportion_age_stratified}
\alias{england_covid_proportion_age_stratified}
\title{#' The England COVID-19 age stratified proportion model dataset}
\usage{
england_covid_proportion_age_stratified()
}
\value{
A dataframe containing the following columns:
\itemize{
\item class (enum(\verb{00_04},\verb{05_09},\verb{10_14},\verb{15_19},\verb{20_24},\verb{25_29},\verb{30_34},\verb{35_39},\verb{40_44},\verb{45_49},\verb{50_54},\verb{55_59},\verb{60_64},\verb{65_69},\verb{70_74},\verb{75_79},\verb{80_84},\verb{85_89},\verb{90+})) - the class column
\item time (as.time_period) - the time column
\item proportion.fit (numeric) - the proportion.fit column
\item proportion.se.fit (numeric) - the proportion.se.fit column
\item proportion.0.025 (numeric) - the proportion.0.025 column
\item proportion.0.05 (numeric) - the proportion.0.05 column
\item proportion.0.25 (numeric) - the proportion.0.25 column
\item proportion.0.5 (numeric) - the proportion.0.5 column
\item proportion.0.75 (numeric) - the proportion.0.75 column
\item proportion.0.95 (numeric) - the proportion.0.95 column
\item proportion.0.975 (numeric) - the proportion.0.975 column
\item relative.growth.fit (numeric) - the relative.growth.fit column
\item relative.growth.se.fit (numeric) - the relative.growth.se.fit column
\item relative.growth.0.025 (numeric) - the relative.growth.0.025 column
\item relative.growth.0.05 (numeric) - the relative.growth.0.05 column
\item relative.growth.0.25 (numeric) - the relative.growth.0.25 column
\item relative.growth.0.5 (numeric) - the relative.growth.0.5 column
\item relative.growth.0.75 (numeric) - the relative.growth.0.75 column
\item relative.growth.0.95 (numeric) - the relative.growth.0.95 column
\item relative.growth.0.975 (numeric) - the relative.growth.0.975 column
}

Minimally grouped by: class (and other groupings allowed).

26790 rows and 20 columns
}
\description{
This is the output of the following estimator, and is here to speed up some
examples. The proportion represented here is the positive tests in this
age group, versus the positive tests in all age groups. It has to be
considered with respect to the overall population.
}
\details{
\if{html}{\out{<div class="sourceCode R">}}\preformatted{england_covid \%>\% ggoutbreak::proportion_locfit_model(window=14)
}\if{html}{\out{</div>}}
}
\concept{datasets}
\keyword{datasets}
