% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{england_ons_infection_survey}
\alias{england_ons_infection_survey}
\title{The england_ons_infection_survey dataset}
\format{
A dataframe containing the following columns:
\itemize{
\item date (date) - the date column
\item geography (character) - the geography column
\item proportion.0.5 (numeric) - the median proportion of people in the region testing
positive for COVID-19
\item proportion.0.025 (numeric) - the lower CI of the proportion of people in the region testing
positive for COVID-19
\item proportion.0.975 (numeric) - the upper CI of the proportion of people in the region testing
positive for COVID-19
\item denom (integer) - the sample size on which this estimate was made (daily rate inferred from
weekly sample sizes.)
\item time (as.time_period) - the time column
}

No mandatory groupings.

No default value.

9820 rows and 7 columns
}
\usage{
data(england_ons_infection_survey)
}
\description{
The COVID-19 ONS infection survey took a random sample of the population
and provides an estimate of the prevalence of COVID-19 that is supposedly
free from ascertainment bias.
}
\details{
The data is available here:
https://www.ons.gov.uk/file?uri=/peoplepopulationandcommunity/healthandsocialcare/conditionsanddiseases/datasets/coronaviruscovid19infectionsurveydata/2023/20230310covid19infectionsurveydatasetsengland.xlsx
}
\keyword{datasets}
