% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aa-examples.R
\name{example_fns}
\alias{example_fns}
\alias{example_england_covid_by_age}
\alias{example_poisson_age_stratified}
\alias{example_poisson_locfit}
\alias{example_proportion_age_stratified}
\alias{example_ganyani_ip}
\alias{example_du_serial}
\alias{example_ip}
\alias{example_bpm}
\alias{example_serial}
\alias{example_poisson_rt}
\alias{example_poisson_growth_rate}
\alias{example_poisson_rt_smooth}
\alias{example_poisson_rt_2class}
\alias{example_delayed_observation}
\title{Example generators}
\usage{
example_england_covid_by_age()

example_poisson_age_stratified()

example_poisson_locfit()

example_proportion_age_stratified()

example_ganyani_ip()

example_du_serial()

example_ip()

example_bpm()

example_serial()

example_poisson_rt()

example_poisson_growth_rate()

example_poisson_rt_smooth()

example_poisson_rt_2class()

example_delayed_observation()
}
\value{
example output of the stated functions, usually a dataframe.
}
\description{
These are a set of internally cached functions to support
examples. They are exported as internal functions so that the examples can
run correctly and cache their output to prevent excessive repetition of the
code examples.
}
\section{Functions}{
\itemize{
\item \code{example_england_covid_by_age()}: Example input format including age stratified COVID-19 data.

\item \code{example_poisson_age_stratified()}: Example output of \code{\link[=poisson_locfit_model]{poisson_locfit_model()}} run on
age stratified COVID-19 data.

\item \code{example_poisson_locfit()}: Example output of \code{\link[=poisson_locfit_model]{poisson_locfit_model()}} run on
un-stratified COVID-19 data.

\item \code{example_proportion_age_stratified()}: Example output of \code{\link[=proportion_locfit_model]{proportion_locfit_model()}}
An England COVID-19 age stratified proportion model dataset. The proportion
represented here is the positive tests in this age group, versus the positive
tests in all age groups, so represents a relative growth of ne age group
versus others.

\item \code{example_ganyani_ip()}: Generation time estimate from \code{\link[=make_gamma_ip]{make_gamma_ip()}}.
This estimate is truncated to make it compatible with \code{EpiEstim}. Take from
Ganyani T, Kremer C, Chen D, Torneri A, Faes C, Wallinga J, Hens N.
Estimating the generation interval for coronavirus disease (COVID-19) based
on symptom onset data, March 2020. Euro Surveill. 2020 Apr;25(17):2000257.
doi: 10.2807/1560-7917.ES.2020.25.17.2000257. PMID: 32372755; PMCID:
PMC7201952.

\item \code{example_du_serial()}: An undjusted serial interval between symptoms \code{\link[=make_resampled_ip]{make_resampled_ip()}}
Z. Du, X. Xu, Y. Wu, L. Wang, B. J. Cowling, and L. A. Meyers, ‘Serial
Interval of COVID-19 among Publicly Reported Confirmed Cases’, Emerg Infect
Dis, vol. 26, no. 6, pp. 1341–1343, Jun. 2020, doi: 10.3201/eid2606.200357.

\item \code{example_ip()}: Example output of \code{\link[=make_gamma_ip]{make_gamma_ip()}}
A test infectivity profile generated from a set of discretised gamma
distributions with parameters mean 5 (95\% CI 4-6) and sd 2 (95\% CI 1.5-2.5).

\item \code{example_bpm()}: Example output of \code{\link[=sim_branching_process]{sim_branching_process()}}
An example of the linelist output of the branching process model simulation.
This is generated using the \code{example_ip()} infectivity profile and
also includes a delay to symptom onset which is a random gamma distributed
quantity with mean of 6 and standard deviation of 2

\item \code{example_serial()}: Example output of \code{\link[=make_resampled_ip]{make_resampled_ip()}}
A serial interval estimated from symptom onset in simulated data including
negative intervals. This serial interval is resampled from the first 1000
patients in the \code{\link[=example_bpm]{example_bpm()}} dataset for whom both infector and infectee
has symptoms. These patients are generated with a symptom delay of mean 6
days and SD 2 from infection (discrete under-dispersed gamma) and an
infectivity profile with mean 5 days and SD 2 as defined in \code{\link[=example_ip]{example_ip()}}
dataset. This serial interval is relevant to the estimation of $R_t$ from
symptomatic case counts in the \code{\link[=example_bpm]{example_bpm()}} dataset but includes negative
times, and cannot be used with \code{EpiEstim}.

\item \code{example_poisson_rt()}: Example output of \code{\link[=sim_poisson_Rt_model]{sim_poisson_Rt_model()}}
An example of the linelist output of the poisson model simulation with defined
$R_t$. This is generated using the \code{example_ip()} infectivity profile

\item \code{example_poisson_growth_rate()}: Example output of \code{\link[=sim_poisson_model]{sim_poisson_model()}}
A simulation dataset determined by a step function of growth rates. This is
useful for demonstrating growth rate estimators.

\item \code{example_poisson_rt_smooth()}: Example output of \code{\link[=sim_poisson_Rt_model]{sim_poisson_Rt_model()}}
Output of a poisson model simulation with a
smooth function for $R_t$ defined as \verb{R(t) = e^(sin(t/80*pi)^4-0.25))}. This
is a relatively unchallenging test data set that should not pose a problem
for smooth estimators.

\item \code{example_poisson_rt_2class()}: Two class example using \code{\link[=sim_poisson_Rt_model]{sim_poisson_Rt_model()}}
Two smooth $R_t$ based incidence timeseries one growing with an time varying Rt
\code{exp(sin(t / 80 * pi)^4 - 0.25)} and the other offset by 10 days:
\code{exp(sin((t - 10) / 80 * pi)^4 - 0.25)}. This is a simple relative growth test

\item \code{example_delayed_observation()}: Example output of \code{\link[=sim_branching_process]{sim_branching_process()}} with
\code{\link[=sim_apply_delay]{sim_apply_delay()}}
This simulates what might be observed in an outbreak if there was on average
a 5 day delay on the reporting of hospital admissions.
The configuration of the outbreak is the same as \code{\link[=example_bpm]{example_bpm()}}, but
this is summary data that describes the whole history of admissions that were
observed, when observed at any given time point. This is a triangular set
of data where the counts are right censored by the observation time.

}}
\examples{
suppressWarnings({
  example_poisson_age_stratified() \%>\% dplyr::glimpse()
  example_ip() \%>\% dplyr::glimpse()
  example_bpm() \%>\% dplyr::glimpse()
  example_serial() \%>\% dplyr::glimpse()
  example_poisson_rt() \%>\% dplyr::glimpse()
  example_poisson_growth_rate() \%>\% dplyr::glimpse()
  example_poisson_rt_smooth() \%>\% dplyr::glimpse()
  example_poisson_rt_2class() \%>\% dplyr::glimpse()
  example_ganyani_ip() \%>\% dplyr::glimpse()
  example_du_serial() \%>\% dplyr::glimpse()
  example_delayed_observation() \%>\% dplyr::glimpse()
})
}
\keyword{internal}
