% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infectivity-profile.R
\name{format_ip}
\alias{format_ip}
\title{Print a summary of an infectivity profile}
\usage{
format_ip(ip = i_empirical_ip)
}
\arguments{
\item{ip}{the infectivity profile to summarise. \code{a0} and \code{a1} columns are
optional if \code{tau} is given.

A dataframe containing the following columns: 
\itemize{
\item boot (anything + default(1)) - a bootstrap identifier
\item probability (proportion) - the probability of new event during this period.
\item a0 (double) - the beginning of the time period (in days)
\item a1 (double) - the end of the time period (in days)
}

Must be grouped by: boot (exactly).

A default value is defined.}
}
\value{
an infectivity profile description
}
\description{
Print a summary of an infectivity profile
}
\examples{
format_ip(ganyani_ip)

}
