% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-gam.R
\name{gam_knots}
\alias{gam_knots}
\title{Derive a set of knot points for a GAM from data}
\usage{
gam_knots(data = i_incidence_data, window, ..., k = NULL)
}
\arguments{
\item{data}{the function will be called with incidence data - a dataframe with columns:
\itemize{
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
}

Any grouping allowed.
}

\item{window}{the spacing between knots}

\item{...}{currently not used}

\item{k}{alternative to \code{window}, if \code{k} is given then the behaviour of the
knots will be similar to the default \code{mgcv::s(...,k=...)} parameter.}
}
\value{
a vector of times (as a numeric)
}
\description{
At the moment this does nothing sophisticated. An mostly equally spaced grid
of knots with gaps at start and end to prevent over-fitting there. In the
future this could look at the number of observations or areas where there is
a lot of change to add in more knots.
}
\examples{
gam_knots(example_poisson_rt(), 14)
gam_knots(example_poisson_rt(), k=10)
}
\concept{models}
