% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-gam.R
\name{gam_poisson_model_fn}
\alias{gam_poisson_model_fn}
\title{Default GAM count model.}
\usage{
gam_poisson_model_fn(window, ..., knots_fn = ~gam_knots(.x, window, ...))
}
\arguments{
\item{window}{controls the knot spacing in the GAM (if the default)}

\item{...}{
  Named arguments passed on to \code{\link[=gam_knots]{gam_knots}}\describe{
    \item{\code{data}}{the function will be called with incidence data - a dataframe with columns:
\itemize{
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
}

Any grouping allowed.
}
\item{\code{k}}{alternative to \code{window}, if \code{k} is given then the behaviour of the
knots will be similar to the default \code{mgcv::s(...,k=...)} parameter.}
\item{\code{...}}{currently not used}
}}

\item{knots_fn}{a function that takes the data as an input and returns
a set of integers as time points for GAM knots, for \code{s(time)} term. The
default here provides a roughly equally spaced grid determined by \code{window},
by a user supplied function could do anything. The input this function is
the raw dataframe of data that will be considered for one model fit. It is
guaranteed to have at least a \code{time} and \code{count} column. It is possible to}
}
\value{
a function suitable as the input for
\code{poisson_gam_model(model_fn = ...)}.
}
\description{
This function configures a very simple poisson count model and using:
}
\details{
\code{count ~ s(time, bs = "cr", k = length(kts))}

The control of knot positions is defined by the \code{knots_fn} and \code{window}
parameters.
}
\examples{
model_fn = gam_poisson_model_fn(14)
fit = model_fn(test_poisson_rt \%>\% dplyr::ungroup())
summary(fit)

}
\concept{models}
