% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ganyani_infectivity_profile}
\alias{ganyani_infectivity_profile}
\title{The ganyani_infectivity_profile dataset}
\format{
A dataframe containing the following columns:
\itemize{
\item tau (integer) - the time column
\item probability (numeric) - the probability column
\item boot (integer) - the boot column
}

Minimally grouped by: boot (and other groupings allowed).

2400 rows and 3 columns
}
\usage{
data(ganyani_infectivity_profile)
}
\description{
The ganyani_infectivity_profile dataset
}
\keyword{datasets}
