% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ganyani_ip}
\alias{ganyani_ip}
\title{A COVID-19 infectivity profile based on an Ganyani et al 2020}
\format{
A dataframe containing the following columns:
\itemize{
\item boot (anything + default(1)) - a bootstrap identifier
\item probability (proportion) - the probability of infection between previous time period until \code{time}
\item tau (double) - the time index this probability relates to (in days)
\item a0 - the beginning of the time period
\item a1 - the end of the time period
}

Grouped by boot (exactly).

A dataframe containing the following columns:
\itemize{
\item boot (anything + default(1)) - a bootstrap identifier
\item probability (proportion) - the probability of infection between previous time period until \code{time}
\item tau (numeric) - the time index this probability relates to (in days)
\item a0 (numeric) - the beginning of the time period
\item a1 (numeric) - the end of the time period
}

Grouped by: boot.

2400 rows and 5 columns
}
\usage{
data(ganyani_ip)
}
\description{
A COVID-19 infectivity profile based on an Ganyani et al 2020

The Ganyani serial interval dataset, compatible with \code{EpiEstim}
}
\references{
\url{https://www.eurosurveillance.org/content/10.2807/1560-7917.ES.2020.25.17.2000257}
}
\concept{datasets}
