% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-common.R
\name{geom_truth}
\alias{geom_truth}
\title{Add a "true" time series to a plot}
\usage{
geom_truth(
  true_df = NULL,
  ...,
  true_col = NULL,
  true_fmt = list(colour = "red")
)
}
\arguments{
\item{true_df}{a data frame with a \code{time_period} column called time and a
value column (name given in \code{true_col}). This is optional and will be picked
up from the \code{raw} parameter if given.}

\item{...}{
  Named arguments passed on to \code{\link[=as.time_period]{as.time_period}}\describe{
    \item{\code{unit}}{the length of one unit of time. This will be either a integer
number of days, or a specification such as "1 week", or another \code{time_period}.
If \code{x} is a \code{time_period}, and the unit is different to that of \code{x}
this will return a rescaled \code{time_period} using the new units.}
\item{\code{start_date}}{the zero time date as something that can be coerced to a
date. If the \code{x} input is already a \code{time_period} and this is different to its
\code{start_date} then \code{x} will be recalibrated to use the new start date.}
\item{\code{anchor}}{only relevant if \code{x} is a vector of dates, this is a date, or
\code{"start"} or \code{"end"} or a weekday name e.g.
\code{"mon"}. With the vector of dates in \code{x} it will use this anchor to find a reference date for
the time-series. If not provided then the current defaults will be used.
(see \code{\link[=set_defaults]{set_defaults()}})}
}}

\item{true_col}{the column name / expression of the true value}

\item{true_fmt}{a list of ggplot formatting to apply to the true value timeseries}
}
\value{
A geom with the true value as a line.
}
\description{
Add a "true" time series to a plot
}
\concept{vis}
\keyword{internal}
