% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggoutbreak-package.R
\docType{package}
\name{ggoutbreak-package}
\alias{ggoutbreak}
\alias{ggoutbreak-package}
\title{ggoutbreak: Estimate Incidence, Proportions and Exponential Growth Rates}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Simple statistical models and visualisations for calculating the incidence, proportion, exponential growth rate, and reproduction number of infectious disease case time series. This toolkit was largely developed during the COVID-19 pandemic.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ai4ci.github.io/ggoutbreak/}
  \item \url{https://github.com/ai4ci/ggoutbreak}
  \item \doi{10.5281/zenodo.13165561}
}

}
\author{
\strong{Maintainer}: Robert Challen \email{rob.challen@bristol.ac.uk} (\href{https://orcid.org/0000-0002-5504-7768}{ORCID}) [copyright holder]

}
\keyword{internal}
