% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot.R
\name{\%above\%}
\alias{\%above\%}
\title{Insert a layer at the bottom of a \code{ggplot}}
\usage{
plot \%above\% layer
}
\arguments{
\item{plot}{the plot to add the layer to}

\item{layer}{the layer to insert underneath the plot}
}
\value{
a \code{ggplot}
}
\description{
Insert a layer at the bottom of a \code{ggplot}
}
\concept{vis}
