% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-sgolay-growth-rate.R
\name{growth_rate_from_incidence}
\alias{growth_rate_from_incidence}
\title{Estimate growth rate from modelled incidence}
\usage{
growth_rate_from_incidence(d = i_incidence_model, window = 7, deg = 2)
}
\arguments{
\item{d}{a modelled incidence estimate - a dataframe with columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item incidence.fit (double) - an estimate of the incidence rate on a log scale
\item incidence.se.fit (positive_double) - the standard error of the incidence rate estimate on a log scale
\item incidence.0.025 (positive_double) - lower confidence limit of the incidence rate (true scale)
\item incidence.0.5 (positive_double) - median estimate of the incidence rate (true scale)
\item incidence.0.975 (positive_double) - upper confidence limit of the incidence rate (true scale)
}

Any grouping allowed.
}

\item{window}{the width of the Savitsky-Golay filter - must be odd}

\item{deg}{the polynomial degree of the filter}
}
\value{
the timeseries with growth rate columns: A dataframe containing the following columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item growth.fit (double) - an estimate of the growth rate
\item growth.se.fit (positive_double) - the standard error the growth rate
\item growth.0.025 (double) - lower confidence limit of the growth rate
\item growth.0.5 (double) - median estimate of the growth rate
\item growth.0.975 (double) - upper confidence limit of the growth rate
}

Any grouping allowed.
}
\description{
This assumes a modelled incidence estimate that is log-normal. The
exponential growth rate is the first derivative of the mu parameters of this
log-normal. On the link scale these are normally distributed. This function
assumes that the time series incidence estimates are uncorrelated to estimate
the error in the growth rate, which is a conservative approach resulting in
more uncertainty in growth rate than might be possible through other methods.
This is all based on Savitsky-Golay filters applied to the normally
distributed log-incidence estimates.
}
\examples{
data = test_poisson_growth_rate
tmp2 = data \%>\%
  poisson_glm_model(window=7,deg=1) \%>\%
  growth_rate_from_incidence(window = 13, deg=1)

if(interactive()) {
  plot_growth_rate(
      tmp2,
      date_labels="\%b \%y"
    )+
    sim_geom_function(data,colour="red")
}
}
\concept{models}
