% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise-proportion.R
\name{infer_rate_ratio}
\alias{infer_rate_ratio}
\title{Calculate a risk ratio from incidence}
\usage{
infer_rate_ratio(
  modelled = i_incidence_model,
  base = i_baseline_incidence_data,
  ...
)
}
\arguments{
\item{modelled}{Model output from something like \code{\link[=poisson_locfit_model]{poisson_locfit_model()}}.
It really makes sense if this is a grouped model. - a dataframe with columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item incidence.fit (double) - an estimate of the incidence rate on a log scale
\item incidence.se.fit (positive_double) - the standard error of the incidence rate estimate on a log scale
\item incidence.0.025 (positive_double) - lower confidence limit of the incidence rate (true scale)
\item incidence.0.5 (positive_double) - median estimate of the incidence rate (true scale)
\item incidence.0.975 (positive_double) - upper confidence limit of the incidence rate (true scale)
}

Any grouping allowed.
}

\item{base}{The baseline data must be grouped in the same way as \code{modelled}.
It may be a time series but does not have to be. See the example and note
this may change in the future. - a dataframe with columns:
\itemize{
\item baseline_incidence (positive_double) - Baseline raw incidence rate as count data
}

Any grouping allowed.
}

\item{...}{not used}
}
\value{
a dataframe with incidence rate ratios for each of the classes in modelled.
A dataframe containing the following columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item rate_ratio.0.025 (positive_double) - lower confidence limit of the rate ratio for a population group
\item rate_ratio.0.5 (positive_double) - median estimate of the rate ratio for a population group
\item rate_ratio.0.975 (positive_double) - upper confidence limit of the rate ratio for a population group
}

Any grouping allowed.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This enables incidence rates are able to be compared to a baseline figure for
incidence. The baseline could come for example from a population average or
average incidence over time. The output is an incidence rate ratio. The
\code{incidence_baseline} column is a rate of events per unit time. The time unit
is expected to be the same as that of the date in \code{modelled} and this is not
checked.
}
\examples{

# not age stratified
baseline = example_poisson_locfit() \%>\%
  dplyr::mutate(baseline_incidence = incidence.0.5)

# age stratified rate ratios:
tmp = example_poisson_age_stratified() \%>\%
  infer_rate_ratio(baseline) \%>\%
  dplyr::glimpse()


}
\concept{models}
