% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise-proportion.R
\name{infer_risk_ratio}
\alias{infer_risk_ratio}
\title{Calculate a normalised risk ratio from proportions}
\usage{
infer_risk_ratio(
  modelled = i_proportion_model,
  base = i_baseline_proportion_data,
  ...
)
}
\arguments{
\item{modelled}{Model output from something like \code{\link[=proportion_locfit_model]{proportion_locfit_model()}} - a dataframe with columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item proportion.fit (double) - an estimate of the proportion on a logit scale
\item proportion.se.fit (positive_double) - the standard error of proportion estimate on a logit scale
\item proportion.0.025 (proportion) - lower confidence limit of proportion (true scale)
\item proportion.0.5 (proportion) - median estimate of proportion (true scale)
\item proportion.0.975 (proportion) - upper confidence limit of proportion (true scale)
}

Any grouping allowed.
}

\item{base}{The baseline data must be grouped in the same way as \code{modelled}.
It may be a time series but does not have to be. - a dataframe with columns:
\itemize{
\item baseline_proportion (proportion) - Baseline proportion for comparison
}

Any grouping allowed.
}

\item{...}{not used}
}
\value{
a dataframe with relative risk / risk ratio columns.
A dataframe containing the following columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item risk_ratio.0.025 (positive_double) - lower confidence limit of the excess risk ratio for a population group
\item risk_ratio.0.5 (positive_double) - median estimate of the excess risk ratio for a population group
\item risk_ratio.0.975 (positive_double) - upper confidence limit of the excess risk ratio for a population group
}

Any grouping allowed.
}
\description{
This assumes that for example, case distribution proportions are stratified
by a population grouping, e.g. geography or age, and we have estimates of the
size of that population during that time period. Normalising by population
proportion allows us to compare the relative risk of the outcome in groups,
compared to the expected population risk if the outcome is evenly distributed
across the population. There may be other proportions other than population
fractions that may be useful to compare. At the moment this does not handle
any uncertainty.
}
\examples{



tmp = ggoutbreak::england_covid \%>\%
  ggoutbreak::proportion_locfit_model(window=21) \%>\%
  ggoutbreak::infer_risk_ratio(ggoutbreak::england_demographics) \%>\%
  dplyr::glimpse()

if(interactive()) {
  plot_growth_phase(tmp)
}

}
\concept{models}
