% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot.R
\name{integer_breaks}
\alias{integer_breaks}
\title{Strictly integer breaks for continuous scale}
\usage{
integer_breaks(n = 5, ...)
}
\arguments{
\item{n}{number of breaks}

\item{...}{further arguments for methods.}
}
\value{
a ggplot breaks function
}
\description{
Strictly integer breaks for continuous scale
}
\concept{vis}
