% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-rt-wallinga.R
\name{inv_wallinga_lipsitch}
\alias{inv_wallinga_lipsitch}
\title{Calculate a growth rate from a reproduction number and an infectivity profile}
\usage{
inv_wallinga_lipsitch(
  Rt,
  y = i_empirical_ip,
  a1 = seq(0.5, length.out = length(y)),
  a0 = dplyr::lag(a1, default = 0)
)
}
\arguments{
\item{Rt}{a vector of reproduction numbers}

\item{y}{an empirical infectivity profile as a probability vector or as a
dataframe of format:
A dataframe containing the following columns:
\itemize{
\item boot (anything + default(1)) - a bootstrap identifier
\item probability (proportion) - the probability of new event during this period.
\item a0 (double) - the beginning of the time period (in days)
\item a1 (double) - the end of the time period (in days)
}

Minimally grouped by: boot (and other groupings allowed).
A default value is defined.}

\item{a1}{the end time of the infectivity profile probability estimate (defaults to 0.5,1.5,2.5,...).}

\item{a0}{the start time of the infectivity profile probability estimate (defaults to 0,0.5,1.5,...).}
}
\value{
an vector of growth rates
}
\description{
Calculate a growth rate from a reproduction number and an infectivity profile
}
\examples{
y=stats::pgamma(seq(0.5,length.out=50), 5,2)-stats::pgamma(c(0,seq(0.5,length.out=49)), 5,2)
inv_wallinga_lipsitch(Rt=seq(0.5,2.5,length.out=9), y=y)
}
\concept{models}
