% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-import.R
\name{linelist}
\alias{linelist}
\title{Coerce an object to a \code{ggoutbreak} compatible case linelist.}
\usage{
linelist(x, ...)
}
\arguments{
\item{x}{An object to coerce}

\item{...}{
  Named arguments passed on to \code{\link[=as.time_period]{as.time_period}}\describe{
    \item{\code{unit}}{the length of one unit of time. This will be either a integer
number of days, or a specification such as "1 week", or another \code{time_period}.
If \code{x} is a \code{time_period}, and the unit is different to that of \code{x}
this will return a rescaled \code{time_period} using the new units.}
\item{\code{start_date}}{the zero time date as something that can be coerced to a
date. If the \code{x} input is already a \code{time_period} and this is different to its
\code{start_date} then \code{x} will be recalibrated to use the new start date.}
\item{\code{anchor}}{only relevant if \code{x} is a vector of dates, this is a date, or
\code{"start"} or \code{"end"} or a weekday name e.g.
\code{"mon"}. With the vector of dates in \code{x} it will use this anchor to find a reference date for
the time-series. If not provided then the current defaults will be used.
(see \code{\link[=set_defaults]{set_defaults()}})}
}}
}
\value{
minimally a time stamped linelist dataframe A dataframe containing the following columns:
\itemize{
\item time (ggoutbreak::time_period) - A set of events with a timestamp as a \code{time_period}
}

Any grouping allowed.
}
\description{
Coerce an object to a \code{ggoutbreak} compatible case linelist.
}
\examples{
(Sys.Date()+stats::runif(100)*7) \%>\% linelist()
}
