% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infectivity-profile.R
\name{make_empirical_ip}
\alias{make_empirical_ip}
\title{Recover a long format infectivity profile from an \code{EpiEstim} style matrix}
\usage{
make_empirical_ip(omega, normalise = TRUE)
}
\arguments{
\item{omega}{a matrix of probabilities, starting at time zero, with columns
representing one possible infectivity profile, with the fist value being
the probability at time zero to 0.5. Or a vector of probabilities for one
single profile, resulting in 1 bootstrap.}

\item{normalise}{is this a probability mass function? In which case we make the
sum of this equal to one (the default). If \code{FALSE} then the input matrix,
or vector is clipped so that its maximum value is one. If this is a number
then the PMF is scaled to this value.}
}
\value{
a long format \code{ip} delay distribution
}
\description{
Recover a long format infectivity profile from an \code{EpiEstim} style matrix
}
\examples{
format_ip(make_empirical_ip(c(0,0,1,1,1,2,2,2,1,1)))
}
\concept{delay_distribution}
