% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infectivity-profile.R
\name{make_fixed_ip}
\alias{make_fixed_ip}
\title{Generate a simple discrete infectivity profile from a gamma distribution}
\usage{
make_fixed_ip(mean, sd = sqrt(mean), epiestim_compat = FALSE)
}
\arguments{
\item{mean}{The mean of a gamma distribution}

\item{sd}{The sd of a gamma distribution}

\item{epiestim_compat}{Should the discretisation support EpiEstim
(i.e. \code{Pr(x<1) = 0})}
}
\value{
a discrete infectivity profile with 1 bootstrap
}
\description{
Generate a simple discrete infectivity profile from a gamma distribution
}
\examples{
tmp = make_fixed_ip(5,2)

if(interactive()) {
  plot_ip(tmp, alpha=1)
}
}
\concept{delay_distribution}
