% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infectivity-profile.R
\name{make_resampled_ip}
\alias{make_resampled_ip}
\title{Re-sample an empirical IP distribution direct from data}
\usage{
make_resampled_ip(
  tau,
  min_tau = pmax(tau - 0.5, truncate),
  max_tau = tau + 0.5,
  add_noise = TRUE,
  truncate = 0,
  n_boots = 100,
  seed = Sys.time()
)
}
\arguments{
\item{tau}{the delay between first and second events}

\item{min_tau}{the minimum delay for interval censored delays}

\item{max_tau}{the maximum delay for interval censored delays}

\item{add_noise}{adds noise to each date point for each replicate}

\item{truncate}{what is the minimum realistic value of the parameter}

\item{n_boots}{number of replicates to generate}

\item{seed}{a random number seed for reproducibility}
}
\value{
a long format \code{ip} delay distribution
}
\description{
Suits larger contact tracing data sets where there is a delay between
2 events which may or may not be precisely known.
}
\examples{
tau = rgamma2(100, 5,2)
ip = make_resampled_ip(min_tau = tau-1, max_tau = tau+1, seed = 100)
if(interactive()) {
  plot_ip(ip,alpha=0.1)
}
}
\concept{delay_distribution}
