% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic-times.R
\name{months.time_period}
\alias{months.time_period}
\title{Extract Parts of a POSIXt or Date Object}
\usage{
\method{months}{time_period}(x, abbreviate = FALSE)
}
\arguments{
\item{x}{an object inheriting from class \code{"POSIXt"} or \code{"Date"}.}

\item{abbreviate}{logical vector (possibly recycled).  Should the names be
    abbreviated?}
}
\value{
\code{weekdays} and \code{months} return a character
  vector of names in the locale in use, i.e., \code{\link[base]{Sys.getlocale}("LC_TIME")}.

  \code{quarters} returns a character vector of \code{"Q1"} to
  \code{"Q4"}.

  \code{julian} returns the number of days (possibly fractional)
  since the origin, with the origin as a \code{"origin"} attribute.
  All time calculations in \R are done ignoring leap-seconds.
}
\description{
Extract the weekday, month or quarter, or the Julian time
  (days since some origin).  These are generic functions: the methods
  for the internal date-time classes are documented here.
}
\note{
Other components such as the day of the month or the year are
  very easy to compute: just use \code{\link[base]{as.POSIXlt}} and extract
  the relevant component.  Alternatively (especially if the components
  are desired as character strings), use \code{\link[base]{strftime}}.
}
\examples{
\donttest{## first two are locale dependent:
weekdays(.leap.seconds)
months  (.leap.seconds)}
quarters(.leap.seconds)

## Show how easily you get month, day, year, day (of {month, week, yr}), ... :
## (remember to count from 0 (!): mon = 0..11, wday = 0..6,  etc !!)

##' Transform (Time-)Date vector  to  convenient data frame :
dt2df <- function(dt, dName = deparse(substitute(dt))) {
    DF <- as.data.frame(unclass(as.POSIXlt( dt )))
    `names<-`(cbind(dt, DF, deparse.level=0L), c(dName, names(DF)))
}
## e.g.,
dt2df(.leap.seconds)    # date+time
\donttest{dt2df(Sys.Date() + 0:9) # date
}
##' Even simpler:  Date -> Matrix - dropping time info {sec,min,hour, isdst}
d2mat <- function(x) simplify2array(unclass(as.POSIXlt(x))[4:7])
## e.g.,
d2mat(seq(as.Date("2000-02-02"), by=1, length.out=30)) # has R 1.0.0's release date

\donttest{
## Julian Day Number (JDN, https://en.wikipedia.org/wiki/Julian_day)
## is the number of days since noon UTC on the first day of 4317 BCE.
## in the proleptic Julian calendar.  To more recently, in
## 'Terrestrial Time' which differs from UTC by a few seconds
## See https://en.wikipedia.org/wiki/Terrestrial_Time
julian(Sys.Date(), -2440588) # from a day
floor(as.numeric(julian(Sys.time())) + 2440587.5) # from a date-time
}
}
\seealso{
\code{\link[base]{DateTimeClasses}}, \code{\link[base]{Date}};
  \code{\link[base]{Sys.getlocale}("LC_TIME")} crucially for \code{months()} and \code{weekdays()}.
}
\concept{time_period}
