% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-multinomial.R
\name{multinomial_nnet_model}
\alias{multinomial_nnet_model}
\title{Multinomial time-series model.}
\usage{
multinomial_nnet_model(
  d = i_multinomial_input,
  ...,
  window = 14,
  frequency = "1 day",
  predict = TRUE,
  .progress = interactive()
)
}
\arguments{
\item{d}{A multi-class count input dataframe - a dataframe with columns:
\itemize{
\item class (factor) - A factor specifying the type of observation. This will be things like
variant, or serotype, for a multinomial model. Any missing data points are ignored.
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
}

Minimally grouped by: class (and other groupings allowed).
}

\item{...}{not used}

\item{window}{a number of data points between knots, smaller values result in
less smoothing, large value in more.}

\item{frequency}{the temporal density of the output estimates.}

\item{predict}{result a prediction. If false we return the model.}

\item{.progress}{show a CLI progress bar}
}
\value{
a new dataframe with \code{time} (as a time period), \code{class}, and \code{proportion.0.5}, or a model object
}
\description{
Takes a list of times, classes and counts, e.g. a COGUK variant like data set
with time, (multinomial) class (e.g. variant) and count being the count in
that time period. Fits a quadratic B-spline on time to the proportion of the
data using \code{nnet::multinom}, with approx one degree of freedom per class and per
\code{window} units of the time series.
}
\details{
Additional groupings are treated as distinct proportions models.
}
\examples{

data = example_poisson_rt_2class()
tmp = data \%>\% multinomial_nnet_model(window=14)

if (interactive()) {
  plot_multinomial(tmp, date_labels="\%b \%y")
}

}
\concept{models}
