% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise-incidence.R
\name{normalise_count}
\alias{normalise_count}
\title{Calculate a normalised count per capita}
\usage{
normalise_count(
  raw = i_incidence_data,
  pop = i_population_data,
  ...,
  population_unit = 1e+05,
  normalise_time = FALSE
)
}
\arguments{
\item{raw}{The count data - a dataframe with columns:
\itemize{
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
}

Any grouping allowed.
}

\item{pop}{The population data must be grouped in the same way as \code{raw}. - a dataframe with columns:
\itemize{
\item population (positive_integer) - Size of population
}

Any grouping allowed.
}

\item{...}{not used}

\item{population_unit}{What population unit do you want the count data
normalised to e.g. per 100K}

\item{normalise_time}{The default behaviour for normalising is to keep it in
the same time units as the input data. If this parameter is set to \code{TRUE}
the incidence rates are calculated per year. If given as a lubridate period
string e.g. "1 week" then the incidence is calculated over that time
period.}
}
\value{
a dataframe with incidence rates per unit capita.
A dataframe containing the following columns:
\itemize{
\item population (positive_integer) - Size of population
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
}

Any grouping allowed.
}
\description{
This assumes positive disease counts are stratified by a population grouping,
e.g. geography or age, and we have estimates of the size of that population
during that time period. Normalising by population size allows us to compare
groups.
}
\examples{

data = example_england_covid_by_age()
demog = ukc19::uk_population_2019_by_5yr_age

data \%>\%
  normalise_count(demog) \%>\%
  dplyr::glimpse()

}
\concept{models}
