% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infectivity-profile.R
\name{omega_matrix}
\alias{omega_matrix}
\title{Generate a infectivity profile matrix from a long format}
\usage{
omega_matrix(ip = i_discrete_ip, epiestim_compat = TRUE)
}
\arguments{
\item{ip}{long format infectivity profile - a dataframe with columns:
\itemize{
\item boot (anything + default(1)) - a bootstrap identifier
\item probability (proportion) - the probability of new event during this period.
\item tau (integer + complete) - the days since the index event.
}

Minimally grouped by: boot (and other groupings allowed).

A default value is defined.
}

\item{epiestim_compat}{ensure the matrix works with \code{EpiEstim}'s
\code{si_from_sample} method by making probability at time zero equal to zero.}
}
\value{
a matrix with \code{0:max(tau)} rows and \code{1:max(boot)} columns.
}
\description{
This converts a long format infectivity profile to something that will work
with \code{EpiEstim}. In general \code{ggoutbreak} will do the conversion internally.
This is provided as a utility for examples mostly.
}
\examples{
omega_matrix(ggoutbreak::covid_ip)
}
\concept{delay_distribution}
