% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{pbeta2}
\alias{pbeta2}
\title{The Beta Distribution}
\source{
\itemize{
  \item The central \code{dbeta} is based on a binomial probability, using code
  contributed by Catherine Loader (see \code{\link[stats]{dbinom}}) if either
  shape parameter is larger than one, otherwise directly from the definition.
  The non-central case is based on the derivation as a Poisson
  mixture of betas (Johnson \abbr{et al.}, 1995, pp.\ifelse{latex}{\out{~}}{ }{}502--3).

  \item The central \code{pbeta} for the default (\code{log_p = FALSE})
  uses a C translation based on

  Didonato, A. and Morris, A., Jr, (1992)
  Algorithm 708: Significant digit computation of the incomplete beta
  function ratios,
  \emph{ACM Transactions on Mathematical Software}, \bold{18}, 360--373,
  \ifelse{text}{doi:10.1145/131766.131776 <https://doi.org/10.1145/131766.131776>}{\ifelse{latex}{\href{https://doi.org/10.1145/131766.131776}{doi:10.1145\out{\\slash{}}131766.131776}}{\href{https://doi.org/10.1145/131766.131776}{doi:10.1145/131766.131776}}}.
  (See also\cr
  Brown, B. and Lawrence Levy, L. (1994)
  Certification of algorithm 708: Significant digit computation of the
  incomplete beta,
  \emph{ACM Transactions on Mathematical Software}, \bold{20}, 393--397,
  \ifelse{text}{doi:10.1145/192115.192155 <https://doi.org/10.1145/192115.192155>}{\ifelse{latex}{\href{https://doi.org/10.1145/192115.192155}{doi:10.1145\out{\\slash{}}192115.192155}}{\href{https://doi.org/10.1145/192115.192155}{doi:10.1145/192115.192155}}}.)
  \cr %%
  We have slightly tweaked the original \dQuote{TOMS 708} algorithm, and
  enhanced for \code{log.p = TRUE}.  For that (log-scale) case,
  underflow to \code{-Inf} (i.e., \eqn{P = 0}) or \code{0}, (i.e.,
  \eqn{P = 1}) still happens because the original algorithm was designed
  without log-scale considerations.  Underflow to \code{-Inf} now
  typically signals a \code{\link{warning}}.

  \item The non-central \code{pbeta} uses a C translation of

  Lenth, R. V. (1987) Algorithm AS 226: Computing noncentral beta
  probabilities. \emph{Applied Statistics}, \bold{36}, 241--244,
  \ifelse{text}{doi:10.2307/2347558 <https://doi.org/10.2307/2347558>}{\ifelse{latex}{\href{https://doi.org/10.2307/2347558}{doi:10.2307\out{\\slash{}}2347558}}{\href{https://doi.org/10.2307/2347558}{doi:10.2307/2347558}}}, 
  incorporating\cr
  Frick, H. (1990)'s AS R84, \emph{Applied Statistics}, \bold{39},
  311--2, \ifelse{text}{doi:10.2307/2347780 <https://doi.org/10.2307/2347780>}{\ifelse{latex}{\href{https://doi.org/10.2307/2347780}{doi:10.2307\out{\\slash{}}2347780}}{\href{https://doi.org/10.2307/2347780}{doi:10.2307/2347780}}}
  and\cr
  Lam, M.L. (1995)'s AS R95, \emph{Applied Statistics}, \bold{44},
  551--2, \ifelse{text}{doi:10.2307/2986147 <https://doi.org/10.2307/2986147>}{\ifelse{latex}{\href{https://doi.org/10.2307/2986147}{doi:10.2307\out{\\slash{}}2986147}}{\href{https://doi.org/10.2307/2986147}{doi:10.2307/2986147}}}.

  This computes the lower tail only, so the upper tail suffers from
  cancellation and a warning will be given when this is likely to be
  significant.

  \item The central case of \code{qbeta} is based on a C translation of

  Cran, G. W., K. J. Martin and G. E. Thomas (1977).
  Remark AS R19 and Algorithm AS 109,
  \emph{Applied Statistics},  \bold{26}, 111--114,
  \ifelse{text}{doi:10.2307/2346887 <https://doi.org/10.2307/2346887>}{\ifelse{latex}{\href{https://doi.org/10.2307/2346887}{doi:10.2307\out{\\slash{}}2346887}}{\href{https://doi.org/10.2307/2346887}{doi:10.2307/2346887}}},
  and subsequent remarks (AS83 and correction).

  Enhancements, notably for starting values and switching to a log-scale
  Newton search, by R Core.

  \item  The central case of \code{rbeta} is based on a C translation of

  R. C. H. Cheng (1978).
  Generating beta variates with nonintegral shape parameters.
  \emph{Communications of the ACM}, \bold{21}, 317--322.
 }
}
\usage{
pbeta2(q, prob, kappa, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{prob}{the mean probability (vectorised)}

\item{kappa}{a coefficient of variation. where 0 is no variability and 1 is
maximally variability (vectorised)}

\item{lower.tail}{logical; if TRUE (default), probabilities are
    \eqn{P[X \le x]}, otherwise, \eqn{P[X > x]}.}
}
\value{
\code{dbeta} gives the density, \code{pbeta} the distribution
  function, \code{qbeta} the quantile function, and \code{rbeta}
  generates random deviates.

  Invalid arguments will result in return value \code{NaN}, with a warning.

  The length of the result is determined by \code{n} for
  \code{rbeta}, and is the maximum of the lengths of the
  numerical arguments for the other functions.

  The numerical arguments other than \code{n} are recycled to the
  length of the result.  Only the first elements of the logical
  arguments are used.
}
\description{
Density, distribution function, quantile function and random
  generation for the Beta distribution with parameters \code{shape1} and
  \code{shape2} (and optional non-centrality parameter \code{ncp}).
}
\details{
The Beta distribution with parameters \code{shape1} \eqn{= a} and
  \code{shape2} \eqn{= b} has density
  \deqn{f(x)=\frac{\Gamma(a+b)}{\Gamma(a)\Gamma(b)}{x}^{a-1} {(1-x)}^{b-1}%
  }{\\Gamma(a+b)/(\\Gamma(a)\\Gamma(b)) x^(a-1) (1-x)^(b-1)}
  for \eqn{a > 0}, \eqn{b > 0} and \eqn{0 \le x \le 1}
  where the boundary values at \eqn{x=0} or \eqn{x=1} are defined as
  by continuity (as limits).
  \cr
  The mean is \eqn{a/(a+b)} and the variance is \eqn{ab/((a+b)^2 (a+b+1))}.
  If \eqn{a,b > 1}, (or one of them \eqn{=1}), the mode is \eqn{(a-1)/(a+b-2)}.
  These and all other distributional properties can be defined as
  limits (leading to point masses at 0, 1/2, or 1) when \eqn{a} or
  \eqn{b} are zero or infinite, and the corresponding
  \code{[dpqr]beta()} functions are defined correspondingly.

  \code{pbeta} is closely related to the incomplete beta function.  As
  defined by Abramowitz and Stegun 6.6.1
  \deqn{B_x(a,b) = \int_0^x t^{a-1} (1-t)^{b-1} dt,}{
    B_x(a,b) = integral_0^x t^(a-1) (1-t)^(b-1) dt,}
  and 6.6.2 \eqn{I_x(a,b) = B_x(a,b) / B(a,b)} where
  \eqn{B(a,b) = B_1(a,b)} is the Beta function (\code{\link{beta}}).

  \eqn{I_x(a,b)} is \code{pbeta(x, a, b)}.

  The noncentral Beta distribution (with \code{ncp} \eqn{ = \lambda})
  is defined (Johnson \abbr{et al.}, 1995, pp.\ifelse{latex}{\out{~}}{ }{}502) as the distribution of
  \eqn{X/(X+Y)} where \eqn{X \sim \chi^2_{2a}(\lambda)}{X ~ chi^2_2a(\\lambda)}
  and \eqn{Y \sim \chi^2_{2b}}{Y ~ chi^2_2b}.
  There, \eqn{\chi^2_n(\lambda)}{chi^2_n(\\lambda)} is the noncentral
  chi-squared distribution with \eqn{n} degrees of freedom and
  non-centrality parameter \eqn{\lambda}, see \link[stats]{Chisquare}.
}
\note{
Supplying \code{ncp = 0} uses the algorithm for the non-central
  distribution, which is not the same algorithm as when \code{ncp} is
  omitted.  This is to give consistent behaviour in extreme cases with
  values of \code{ncp} very near zero.
}
\examples{
pbeta2(c(0.25,0.5,0.75), 0.5, 0.25)
}
\references{
Becker, R. A., Chambers, J. M. and Wilks, A. R. (1988)
  \emph{The New S Language}.
  Wadsworth & Brooks/Cole.

  Abramowitz, M. and Stegun, I. A. (1972)
  \emph{Handbook of Mathematical Functions.} New York: Dover.
  Chapter 6: Gamma and Related Functions.

  Johnson, N. L., Kotz, S. and Balakrishnan, N. (1995)
  \emph{Continuous Univariate Distributions}, volume 2, especially
  chapter 25. Wiley, New York.
}
\seealso{
\link[stats]{Distributions} for other standard distributions.

  \code{\link{beta}} for the Beta function.
}
\concept{distributions}
