% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-raw-ts.R
\name{plot_counts}
\alias{plot_counts}
\title{Plot a raw case count timeseries}
\usage{
plot_counts(raw, ..., mapping = .check_for_aes(raw, ...), events = i_events)
}
\arguments{
\item{raw}{The raw count data, or the raw count data normalised by population (
see \code{\link[=normalise_count]{normalise_count()}}) - EITHER: a dataframe with columns:
\itemize{
\item population (positive_integer) - Size of population
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item population_unit (double) - The population unit on which the per capita incidence rate is calculated
\item time_unit (lubridate::as.period) - The time period over which the per capita incidence rate is calculated
}

Any grouping allowed.

OR with columns:
\itemize{
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
}

Any grouping allowed.

}

\item{...}{
  Named arguments passed on to \code{\link[=geom_events]{geom_events}}\describe{
    \item{\code{events}}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}}

\item{mapping}{a \code{ggplot2::aes} mapping. Most importantly setting the \code{colour}
to something if there are multiple incidence timeseries in the plot}

\item{events}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}
\value{
a ggplot object
}
\description{
Plot a raw case count timeseries
}
\examples{
# example code

tmp = example_england_covid_by_age() \%>\%
  time_aggregate(count=sum(count)) \%>\%
  normalise_count(pop=56489700, population_unit=1000, normalise_time=TRUE)

# normalised by England population (56489700 people)

if(interactive()) {
  plot_counts(tmp, colour="blue",size=0.25)
}

}
\concept{vis}
