% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-growth-phase.R
\name{plot_growth_phase}
\alias{plot_growth_phase}
\title{Plot an incidence or proportion versus growth phase diagram}
\usage{
plot_growth_phase(
  modelled,
  timepoints = NULL,
  duration = max(dplyr::count(modelled)$n),
  interval = 7,
  mapping = if (interfacer::is_col_present(modelled, class)) {
     ggplot2::aes(colour =
    class)
 } else {
     ggplot2::aes()
 },
  cis = TRUE,
  ...
)
}
\arguments{
\item{modelled}{Growth rates and incidence / proportion timeseries as the
outputs of functions such as \code{proportion_locfit_model}, \code{poisson_locfit_model},
or similar. - EITHER: a dataframe with columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item incidence.per_capita.fit (double) - an estimate of the incidence per capita rate on a log scale
\item incidence.per_capita.se.fit (positive_double) - the standard error of the incidence per capita rate estimate on a log scale
\item incidence.per_capita.0.025 (positive_double) - lower confidence limit of the incidence per capita rate (true scale)
\item incidence.per_capita.0.5 (positive_double) - median estimate of the incidence per capita rate (true scale)
\item incidence.per_capita.0.975 (positive_double) - upper confidence limit of the incidence per capita rate (true scale)
\item population_unit (double) - The population unit on which the per capita incidence rate is calculated
\item time_unit (lubridate::as.period) - The time period over which the per capita incidence rate is calculated
\item growth.fit (double) - an estimate of the growth rate
\item growth.se.fit (positive_double) - the standard error the growth rate
\item growth.0.025 (double) - lower confidence limit of the growth rate
\item growth.0.5 (double) - median estimate of the growth rate
\item growth.0.975 (double) - upper confidence limit of the growth rate
}

Any grouping allowed.

OR with columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item incidence.fit (double) - an estimate of the incidence rate on a log scale
\item incidence.se.fit (positive_double) - the standard error of the incidence rate estimate on a log scale
\item incidence.0.025 (positive_double) - lower confidence limit of the incidence rate (true scale)
\item incidence.0.5 (positive_double) - median estimate of the incidence rate (true scale)
\item incidence.0.975 (positive_double) - upper confidence limit of the incidence rate (true scale)
\item growth.fit (double) - an estimate of the growth rate
\item growth.se.fit (positive_double) - the standard error the growth rate
\item growth.0.025 (double) - lower confidence limit of the growth rate
\item growth.0.5 (double) - median estimate of the growth rate
\item growth.0.975 (double) - upper confidence limit of the growth rate
}

Any grouping allowed.

OR with columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item risk_ratio.0.025 (positive_double) - lower confidence limit of the excess risk ratio for a population group
\item risk_ratio.0.5 (positive_double) - median estimate of the excess risk ratio for a population group
\item risk_ratio.0.975 (positive_double) - upper confidence limit of the excess risk ratio for a population group
\item proportion.fit (double) - an estimate of the proportion on a logit scale
\item proportion.se.fit (positive_double) - the standard error of proportion estimate on a logit scale
\item proportion.0.025 (proportion) - lower confidence limit of proportion (true scale)
\item proportion.0.5 (proportion) - median estimate of proportion (true scale)
\item proportion.0.975 (proportion) - upper confidence limit of proportion (true scale)
\item relative.growth.fit (double) - an estimate of the relative growth rate
\item relative.growth.se.fit (positive_double) - the standard error the relative growth rate
\item relative.growth.0.025 (double) - lower confidence limit of the relative growth rate
\item relative.growth.0.5 (double) - median estimate of the relative growth rate
\item relative.growth.0.975 (double) - upper confidence limit of the relative growth rate
}

Any grouping allowed.

OR with columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item proportion.fit (double) - an estimate of the proportion on a logit scale
\item proportion.se.fit (positive_double) - the standard error of proportion estimate on a logit scale
\item proportion.0.025 (proportion) - lower confidence limit of proportion (true scale)
\item proportion.0.5 (proportion) - median estimate of proportion (true scale)
\item proportion.0.975 (proportion) - upper confidence limit of proportion (true scale)
\item relative.growth.fit (double) - an estimate of the relative growth rate
\item relative.growth.se.fit (positive_double) - the standard error the relative growth rate
\item relative.growth.0.025 (double) - lower confidence limit of the relative growth rate
\item relative.growth.0.5 (double) - median estimate of the relative growth rate
\item relative.growth.0.975 (double) - upper confidence limit of the relative growth rate
}

Any grouping allowed.

OR with NULL

}

\item{timepoints}{time points (as \code{Date} or \code{time_period} vector) of dates to
plot phase diagrams. If multiple this will result in a sequence of plots
as facets. If \code{NULL} (the default) it will be the last time point in the series}

\item{duration}{the length of the growth rate phase trail}

\item{interval}{the length of time between markers on the phase plot}

\item{mapping}{a \code{ggplot2::aes()} mapping}

\item{cis}{logical; should the phases be marked with confidence intervals?}

\item{...}{
  Named arguments passed on to \code{\link[=geom_events]{geom_events}}\describe{
    \item{\code{events}}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}
  Named arguments passed on to \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap}}\describe{
    \item{\code{facets}}{A set of variables or expressions quoted by \code{\link[ggplot2:vars]{vars()}}
and defining faceting groups on the rows or columns dimension.
The variables can be named (the names are passed to \code{labeller}).

For compatibility with the classic interface, can also be a
formula or character vector. Use either a one sided formula, \code{~a + b},
or a character vector, \code{c("a", "b")}.}
\item{\code{nrow,ncol}}{Number of rows and columns.}
\item{\code{scales}}{Should scales be fixed (\code{"fixed"}, the default),
free (\code{"free"}), or free in one dimension (\code{"free_x"},
\code{"free_y"})?}
\item{\code{space}}{If \code{"fixed"} (default), all panels have the same size and
the number of rows and columns in the layout can be arbitrary. If
\code{"free_x"}, panels have widths proportional to the length of the x-scale,
but the layout is constrained to one row. If \code{"free_y"}, panels have
heights proportional to the length of the y-scale, but the layout is
constrained to one column.}
\item{\code{shrink}}{If \code{TRUE}, will shrink scales to fit output of
statistics, not raw data. If \code{FALSE}, will be range of raw data
before statistical summary.}
\item{\code{labeller}}{A function that takes one data frame of labels and
returns a list or data frame of character vectors. Each input
column corresponds to one factor. Thus there will be more than
one with \code{vars(cyl, am)}. Each output
column gets displayed as one separate line in the strip
label. This function should inherit from the "labeller" S3 class
for compatibility with \code{\link[ggplot2:labeller]{labeller()}}. You can use different labeling
functions for different kind of labels, for example use \code{\link[ggplot2:label_parsed]{label_parsed()}} for
formatting facet labels. \code{\link[ggplot2:label_value]{label_value()}} is used by default,
check it for more details and pointers to other options.}
\item{\code{as.table}}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}} The \code{as.table} argument
is now absorbed into the \code{dir} argument via the two letter options.
If \code{TRUE}, the facets are laid out like a table with highest values at the
bottom-right. If \code{FALSE}, the facets are laid out like a plot with the
highest value at the top-right.}
\item{\code{switch}}{By default, the labels are displayed on the top and
right of the plot. If \code{"x"}, the top labels will be
displayed to the bottom. If \code{"y"}, the right-hand side
labels will be displayed to the left. Can also be set to
\code{"both"}.}
\item{\code{drop}}{If \code{TRUE}, the default, all factor levels not used in the
data will automatically be dropped. If \code{FALSE}, all factor levels
will be shown, regardless of whether or not they appear in the data.}
\item{\code{dir}}{Direction: either \code{"h"} for horizontal, the default, or \code{"v"},
for vertical. When \code{"h"} or \code{"v"} will be combined with \code{as.table} to
set final layout. Alternatively, a combination of \code{"t"} (top) or
\code{"b"} (bottom) with \code{"l"} (left) or \code{"r"} (right) to set a layout directly.
These two letters give the starting position and the first letter gives
the growing direction. For example \code{"rt"} will place the first panel in
the top-right and starts filling in panels right-to-left.}
\item{\code{strip.position}}{By default, the labels are displayed on the top of
the plot. Using \code{strip.position} it is possible to place the labels on
either of the four sides by setting \code{strip.position = c("top",
  "bottom", "left", "right")}}
\item{\code{axes}}{Determines which axes will be drawn in case of fixed scales.
When \code{"margins"} (default), axes will be drawn at the exterior margins.
\code{"all_x"} and \code{"all_y"} will draw the respective axes at the interior
panels too, whereas \code{"all"} will draw all axes at all panels.}
\item{\code{axis.labels}}{Determines whether to draw labels for interior axes when
the scale is fixed and the \code{axis} argument is not \code{"margins"}. When
\code{"all"} (default), all interior axes get labels. When \code{"margins"}, only
the exterior axes get labels, and the interior axes get none. When
\code{"all_x"} or \code{"all_y"}, only draws the labels at the interior axes in the
x- or y-direction respectively.}
}}
}
\value{
a ggplot
}
\description{
Plot an incidence or proportion versus growth phase diagram
}
\examples{

data = ggoutbreak::test_poisson_rt_2class
tmp2 = data \%>\% poisson_locfit_model()

timepoints = as.Date(tmp2$time[c(40,80,120,160)])

if(interactive()) {
  plot_growth_phase(tmp2, timepoints, duration=108)
}
}
\concept{vis}
