% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-growth-rate-ts.R
\name{plot_growth_rate}
\alias{plot_growth_rate}
\title{Growth rate timeseries diagram}
\usage{
plot_growth_rate(
  modelled,
  ...,
  mapping = .check_for_aes(modelled, ...),
  events = i_events
)
}
\arguments{
\item{modelled}{a growth rate dataframe. - EITHER: a dataframe with columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item incidence.fit (double) - an estimate of the incidence rate on a log scale
\item incidence.se.fit (positive_double) - the standard error of the incidence rate estimate on a log scale
\item incidence.0.025 (positive_double) - lower confidence limit of the incidence rate (true scale)
\item incidence.0.5 (positive_double) - median estimate of the incidence rate (true scale)
\item incidence.0.975 (positive_double) - upper confidence limit of the incidence rate (true scale)
\item growth.fit (double) - an estimate of the growth rate
\item growth.se.fit (positive_double) - the standard error the growth rate
\item growth.0.025 (double) - lower confidence limit of the growth rate
\item growth.0.5 (double) - median estimate of the growth rate
\item growth.0.975 (double) - upper confidence limit of the growth rate
}

Any grouping allowed.

OR with columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item proportion.fit (double) - an estimate of the proportion on a logit scale
\item proportion.se.fit (positive_double) - the standard error of proportion estimate on a logit scale
\item proportion.0.025 (proportion) - lower confidence limit of proportion (true scale)
\item proportion.0.5 (proportion) - median estimate of proportion (true scale)
\item proportion.0.975 (proportion) - upper confidence limit of proportion (true scale)
\item relative.growth.fit (double) - an estimate of the relative growth rate
\item relative.growth.se.fit (positive_double) - the standard error the relative growth rate
\item relative.growth.0.025 (double) - lower confidence limit of the relative growth rate
\item relative.growth.0.5 (double) - median estimate of the relative growth rate
\item relative.growth.0.975 (double) - upper confidence limit of the relative growth rate
}

Any grouping allowed.

}

\item{...}{
  Named arguments passed on to \code{\link[=geom_events]{geom_events}}\describe{
    \item{\code{events}}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}}

\item{mapping}{a \code{ggplot2::aes} mapping. Most importantly setting the \code{colour}
to something if there are multiple incidence time series in the plot}

\item{events}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}
\value{
a ggplot
}
\description{
Growth rate timeseries diagram
}
\examples{

data = ggoutbreak::test_poisson_rt_2class
tmp2 = data \%>\% poisson_locfit_model()

if(interactive()) {
  plot_growth_rate(tmp2)
}

}
\concept{vis}
