% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-incidence-ts.R
\name{plot_incidence}
\alias{plot_incidence}
\title{Plot an incidence timeseries}
\usage{
plot_incidence(
  modelled,
  raw = i_incidence_data,
  ...,
  mapping = .check_for_aes(modelled, ...),
  events = i_events
)
}
\arguments{
\item{modelled}{An optional estimate of the incidence time series. If \code{modelled} is
missing then it is estimated from \code{raw} using a \code{poisson_locfit_model}.
In this case parameters \code{window} and \code{deg} may be supplied to control the
fit. \code{modelled} can also be the output from \code{normalise_incidence} in which case
the plot uses the per capita rates calculated by that function. - EITHER: a dataframe with columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item incidence.per_capita.fit (double) - an estimate of the incidence per capita rate on a log scale
\item incidence.per_capita.se.fit (positive_double) - the standard error of the incidence per capita rate estimate on a log scale
\item incidence.per_capita.0.025 (positive_double) - lower confidence limit of the incidence per capita rate (true scale)
\item incidence.per_capita.0.5 (positive_double) - median estimate of the incidence per capita rate (true scale)
\item incidence.per_capita.0.975 (positive_double) - upper confidence limit of the incidence per capita rate (true scale)
\item population_unit (double) - The population unit on which the per capita incidence rate is calculated
\item time_unit (lubridate::as.period) - The time period over which the per capita incidence rate is calculated
}

Any grouping allowed.

OR with columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item incidence.fit (double) - an estimate of the incidence rate on a log scale
\item incidence.se.fit (positive_double) - the standard error of the incidence rate estimate on a log scale
\item incidence.0.025 (positive_double) - lower confidence limit of the incidence rate (true scale)
\item incidence.0.5 (positive_double) - median estimate of the incidence rate (true scale)
\item incidence.0.975 (positive_double) - upper confidence limit of the incidence rate (true scale)
}

Any grouping allowed.

}

\item{raw}{The raw count data (optional - if given overlaid as points on
top of modelled estimate) - a dataframe with columns:
\itemize{
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
}

Any grouping allowed.
}

\item{...}{
  Named arguments passed on to \code{\link[=geom_events]{geom_events}}\describe{
    \item{\code{events}}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}
  Named arguments passed on to \code{\link[=poisson_locfit_model]{poisson_locfit_model}}\describe{
    \item{\code{window}}{a number of data points defining the bandwidth of the estimate,
smaller values result in less smoothing, large value in more. The default
value of 14 is calibrated for data provided on a daily frequency, with
weekly data a lower value may be preferred. - default \code{14}}
\item{\code{deg}}{polynomial degree (min 1) - higher degree results in less
smoothing, lower values result in more smoothing. A degree of 1 is fitting
a linear model piece wise. - default \code{2}}
\item{\code{frequency}}{the density of the output estimates as a time period such as
\verb{7 days} or \verb{2 weeks}. - default \code{1 day}}
}
  Named arguments passed on to \code{\link[=geom_truth]{geom_truth}}\describe{
    \item{\code{true_df}}{a data frame with a \code{time_period} column called time and a
value column (name given in \code{true_col}). This is optional and will be picked
up from the \code{raw} parameter if given.}
\item{\code{true_col}}{the column name / expression of the true value}
\item{\code{true_fmt}}{a list of ggplot formatting to apply to the true value timeseries}
\item{\code{...}}{
  Named arguments passed on to \code{\link[=as.time_period]{as.time_period}}\describe{
    \item{\code{unit}}{the length of one unit of time. This will be either a integer
number of days, or a specification such as "1 week", or another \code{time_period}.
If \code{x} is a \code{time_period}, and the unit is different to that of \code{x}
this will return a rescaled \code{time_period} using the new units.}
\item{\code{start_date}}{the zero time date as something that can be coerced to a
date. If the \code{x} input is already a \code{time_period} and this is different to its
\code{start_date} then \code{x} will be recalibrated to use the new start date.}
\item{\code{anchor}}{only relevant if \code{x} is a vector of dates, this is a date, or
\code{"start"} or \code{"end"} or a weekday name e.g.
\code{"mon"}. With the vector of dates in \code{x} it will use this anchor to find a reference date for
the time-series. If not provided then the current defaults will be used.
(see \code{\link[=set_defaults]{set_defaults()}})}
}}
}}

\item{mapping}{a \code{ggplot2::aes} mapping. Most importantly setting the \code{colour}
to something if there are multiple incidence timeseries in the plot}

\item{events}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}
\value{
a ggplot object
}
\description{
Plot an incidence timeseries
}
\examples{
# example code

tmp = example_poisson_rt_2class()
tmp2 = tmp \%>\% poisson_locfit_model()

if(interactive()) {
  plot_incidence(tmp2,tmp,size=0.25)
}

}
\concept{vis}
