% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-infectivity-profile.R
\name{plot_ip}
\alias{plot_ip}
\title{Plot an infectivity profile}
\usage{
plot_ip(ip = i_empirical_ip, alpha = 0.1, ...)
}
\arguments{
\item{ip}{A long format infectivity profile - a dataframe with columns:
\itemize{
\item boot (anything + default(1)) - a bootstrap identifier
\item probability (proportion) - the probability of new event during this period.
\item a0 (double) - the beginning of the time period (in days)
\item a1 (double) - the end of the time period (in days)
}

Minimally grouped by: boot (and other groupings allowed).

A default value is defined.
}

\item{alpha}{the alpha value of the bootstrap lines}

\item{...}{passed onto geom_segment for controlling line thickness, alpha etc.}
}
\value{
an ggplot object
}
\description{
Plot an infectivity profile
}
\examples{
if(interactive()) {
  plot_ip(ganyani_ip,alpha=0.1)
}
}
\concept{vis}
