% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-multinomial-ts.R
\name{plot_multinomial}
\alias{plot_multinomial}
\title{Plot a multinomial proportions model}
\usage{
plot_multinomial(
  modelled = i_multinomial_proportion_model,
  ...,
  mapping = ggplot2::aes(fill = class),
  events = i_events,
  normalise = FALSE
)
}
\arguments{
\item{modelled}{the multinomial count data - a dataframe with columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item class (factor) - A factor specifying the type of observation. This will be things like
variant, or serotype, for a multinomial model. Any missing data points are ignored.
\item proportion.0.5 (proportion) - median estimate of proportion (true scale)
}

Must be grouped by: class (exactly).
}

\item{...}{
  Named arguments passed on to \code{\link[=geom_events]{geom_events}}\describe{
    \item{\code{events}}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}}

\item{mapping}{a \code{ggplot2::aes} mapping. Usually this will be left as the
default}

\item{events}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}

\item{normalise}{make sure the probabilities add up to one - this can be a
bad idea if you know you may have missing values, on the other hand not all
proportions models are guaranteed to add up to one.}
}
\value{
a ggplot
}
\description{
A multinomial proportions model will tell you what proportion each class has
versus others in the data set. In this case the denominator is the total
count across across all classes.
}
\examples{
tmp = ggoutbreak::england_covid_proportion_age_stratified() \%>\%
  dplyr::glimpse()

if(interactive()) {
  plot_multinomial(tmp, normalise=TRUE)+
    ggplot2::scale_fill_viridis_d()
}

}
\concept{vis}
