% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-prevalence-ts.R
\name{plot_prevalence}
\alias{plot_prevalence}
\title{Plot a proportions timeseries}
\usage{
plot_prevalence(
  modelled = i_prevalence_model,
  raw = i_proportion_data,
  ...,
  mapping = .check_for_aes(modelled, ...),
  events = i_events
)
}
\arguments{
\item{modelled}{Proportion model estimates - a dataframe with columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item prevalence.0.025 (proportion) - lower confidence limit of prevalence (true scale)
\item prevalence.0.5 (proportion) - median estimate of prevalence (true scale)
\item prevalence.0.975 (proportion) - upper confidence limit of prevalence (true scale)
}

Any grouping allowed.
}

\item{raw}{Raw count data - a dataframe with columns:
\itemize{
\item denom (positive_integer) - Total test counts associated with the specified time frame
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
}

Any grouping allowed.
}

\item{...}{
  Named arguments passed on to \code{\link[=geom_events]{geom_events}}\describe{
    \item{\code{events}}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}}

\item{mapping}{a \code{ggplot2::aes} mapping. Most importantly setting the \code{colour}
to something if there are multiple incidence timeseries in the plot}

\item{events}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}
\value{
a ggplot object
}
\description{
Plot a proportions timeseries
}
\examples{
if(interactive()) {
  plot_prevalence(
    ggoutbreak::england_ons_infection_survey,
    mapping = ggplot2::aes(colour=geography))
}
}
\concept{vis}
