% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rt-ts.R
\name{plot_rt}
\alias{plot_rt}
\title{Reproduction number timeseries diagram}
\usage{
plot_rt(
  modelled = i_reproduction_number,
  ...,
  mapping = .check_for_aes(modelled, ...),
  events = i_events
)
}
\arguments{
\item{modelled}{the modelled Rt estimate - a dataframe with columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item rt.fit (double) - an estimate of the reproduction number
\item rt.se.fit (positive_double) - the standard error of the reproduction number
\item rt.0.025 (double) - lower confidence limit of the reproduction number
\item rt.0.5 (double) - median estimate of the reproduction number
\item rt.0.975 (double) - upper confidence limit of the reproduction number
}

Any grouping allowed.
}

\item{...}{
  Named arguments passed on to \code{\link[=geom_events]{geom_events}}\describe{
    \item{\code{events}}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}}

\item{mapping}{a \code{ggplot2::aes} mapping. Most importantly setting the \code{colour}
to something if there are multiple incidence time series in the plot}

\item{events}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}
\value{
a ggplot timeseries
}
\description{
Reproduction number timeseries diagram
}
\examples{
# example code
if (interactive()) {

  tmp2 = england_covid_poisson \%>\%
    rt_from_incidence()

  # comparing RT from growth rates with England consensus Rt
  # (N.B. offset by 17 days to align with estimates):

  plot_rt(tmp2,colour="blue")+
    ggplot2::geom_errorbar(
      data=england_consensus_rt,
      mapping=ggplot2::aes(x=date-17,ymin=low,ymax=high),
      colour="red")

}
}
\concept{vis}
