% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-glm.R
\name{poisson_glm_model}
\alias{poisson_glm_model}
\title{Poisson time-series model.}
\usage{
poisson_glm_model(
  d = i_incidence_input,
  ...,
  window = 14,
  frequency = "1 day",
  .progress = interactive()
)
}
\arguments{
\item{d}{Count model input - a dataframe with columns:
\itemize{
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
}

Any grouping allowed.
}

\item{...}{not used and present to allow proportion model to be used in a
\code{group_modify}}

\item{window}{a number of data points defining the bandwidth of the estimate,
smaller values result in less smoothing, large value in more. The default
value of 14 is calibrated for data provided on a daily frequency, with
weekly data a lower value may be preferred. - default \code{14}}

\item{frequency}{the density of the output estimates as a time period such as
\verb{7 days} or \verb{2 weeks}. - default \code{1 day}}

\item{.progress}{show a CLI progress bar}
}
\value{
A dataframe containing the following columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item incidence.fit (double) - an estimate of the incidence rate on a log scale
\item incidence.se.fit (positive_double) - the standard error of the incidence rate estimate on a log scale
\item incidence.0.025 (positive_double) - lower confidence limit of the incidence rate (true scale)
\item incidence.0.5 (positive_double) - median estimate of the incidence rate (true scale)
\item incidence.0.975 (positive_double) - upper confidence limit of the incidence rate (true scale)
}

Any grouping allowed.
}
\description{
This uses a generalised linear model to fit a quasi-poisson model with a time
varying rate as a natural cubic spline with approx one degree of freedom per
\code{window} units of the time series.
}
\examples{
data = example_poisson_growth_rate()

tmp2 = data \%>\% poisson_glm_model(window=7,deg=2)
tmp3 = data \%>\% poisson_glm_model(window=14,deg=1)

comp = dplyr::bind_rows(
  tmp2 \%>\% dplyr::mutate(class="7:2"),
  tmp3 \%>\% dplyr::mutate(class="14:1"),
) \%>\% dplyr::group_by(class)

if (interactive()) {
  plot_incidence(comp, date_labels="\%b \%y", raw=data, true_col=rate)
}
}
\concept{models}
