% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-glm.R
\name{proportion_glm_model}
\alias{proportion_glm_model}
\title{Binomial time-series model.}
\usage{
proportion_glm_model(
  d = i_proportion_input,
  ...,
  window = 14,
  frequency = "1 day",
  .progress = interactive()
)
}
\arguments{
\item{d}{Proportion model input - a dataframe with columns:
\itemize{
\item denom (positive_integer) - Total test counts associated with the specified time frame
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
}

Any grouping allowed.
}

\item{...}{not used and present to allow proportion model to be used in a
\code{group_modify}}

\item{window}{a number of data points defining the bandwidth of the estimate,
smaller values result in less smoothing, large value in more. The default
value of 14 is calibrated for data provided on a daily frequency, with
weekly data a lower value may be preferred. - default \code{14}}

\item{frequency}{the density of the output estimates as a time period such as
\verb{7 days} or \verb{2 weeks}. - default \code{1 day}}

\item{.progress}{show a CLI progress bar}
}
\value{
A dataframe containing the following columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item proportion.fit (double) - an estimate of the proportion on a logit scale
\item proportion.se.fit (positive_double) - the standard error of proportion estimate on a logit scale
\item proportion.0.025 (proportion) - lower confidence limit of proportion (true scale)
\item proportion.0.5 (proportion) - median estimate of proportion (true scale)
\item proportion.0.975 (proportion) - upper confidence limit of proportion (true scale)
}

Any grouping allowed.
}
\description{
This uses a generalised linear model to fit a quasi-binomial model with a time
varying rate as a natural cubic spline with approx one degree of freedom per
\code{window} units of the time series.
}
\examples{

data = example_poisson_rt_2class()

tmp2 = data \%>\% proportion_glm_model(window=7,deg=2)
tmp3 = data \%>\% proportion_glm_model(window=14,deg=1)

comp = dplyr::bind_rows(
  tmp2 \%>\% dplyr::mutate(model="7:2"),
  tmp3 \%>\% dplyr::mutate(model="14:1"),
) \%>\% dplyr::group_by(model,class)

if (interactive()) {
  plot_proportion(
      comp,
      date_labels="\%b \%y",
      mapping=ggplot2::aes(colour=model),
      raw=data
   )+
   ggplot2::facet_wrap(~class)
}

# TODO: deal with error conditions
# "observations with zero weight not used for calculating dispersion
}
\concept{models}
