% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-locfit.R
\name{proportion_locfit_model}
\alias{proportion_locfit_model}
\title{A binomial proportion estimate and associated exponential growth rate}
\usage{
proportion_locfit_model(
  d = i_proportion_input,
  ...,
  window = 14,
  deg = 2,
  frequency = "1 day",
  predict = TRUE,
  .progress = interactive()
)
}
\arguments{
\item{d}{the input - a dataframe with columns:
\itemize{
\item denom (positive_integer) - Total test counts associated with the specified time frame
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
}

Any grouping allowed.
}

\item{...}{not used and present to allow proportion model to be used in a
\code{group_modify}}

\item{window}{a number of data points defining the bandwidth of the estimate,
smaller values result in less smoothing, large value in more. The default
value of 14 is calibrated for data provided on a daily frequency, with
weekly data a lower value may be preferred. - default \code{14}}

\item{deg}{polynomial degree (min 1) - higher degree results in less
smoothing, lower values result in more smoothing. A degree of 1 is fitting
a linear model piece wise. - default \code{2}}

\item{frequency}{the density of the output estimates as a time period such as
\verb{7 days} or \verb{2 weeks}. - default \code{1 day}}

\item{predict}{result is a prediction dataframe. If false we return the
\code{locfit} models (advanced). - default \code{TRUE}}

\item{.progress}{show a CLI progress bar}
}
\value{
A dataframe containing the following columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item proportion.fit (double) - an estimate of the proportion on a logit scale
\item proportion.se.fit (positive_double) - the standard error of proportion estimate on a logit scale
\item proportion.0.025 (proportion) - lower confidence limit of proportion (true scale)
\item proportion.0.5 (proportion) - median estimate of proportion (true scale)
\item proportion.0.975 (proportion) - upper confidence limit of proportion (true scale)
\item relative.growth.fit (double) - an estimate of the relative growth rate
\item relative.growth.se.fit (positive_double) - the standard error the relative growth rate
\item relative.growth.0.025 (double) - lower confidence limit of the relative growth rate
\item relative.growth.0.5 (double) - median estimate of the relative growth rate
\item relative.growth.0.975 (double) - upper confidence limit of the relative growth rate
}

Any grouping allowed.
}
\description{
takes a list of times, counts and a denominator and fits a quasi-binomial model
using a logit link function to proportion data using local regression
using the package \code{locfit}.
}
\details{
This expects d to contain one combination of:
\itemize{
\item \code{time} and \code{count} and \code{denom} columns - e.g. all tests conducted.
}

This results is a one versus others comparison binomial proportion estimate plus a
relative growth rate estimate specifying how much quicker this is growing
compared to the growth of the denominator.

The denominator maybe the sum of all subgroups \code{denom = sum(count)}, e.g. in
the situation where there are multiple variants of a disease circulating. In
which case the relative growth is that of the subgroup compared to the
overall. You can make this a one-versus-others comparison by making the
denominator exclude the current item (e.g. \code{denom = sum(count)-count}).

The denominator can also be used to express the size of the population tested.
This gives us a relative growth rate that is different in essence to the previous
and may be a better estimate of the true growth rate in the situation where
testing effort is variable, or capacity saturated.
}
\examples{

data = test_poisson_rt_2class

tmp2 = data \%>\% ggoutbreak::proportion_locfit_model(window=7,deg=2)
tmp3 = data \%>\% ggoutbreak::proportion_locfit_model(window=14,deg=1)

comp = dplyr::bind_rows(
  tmp2 \%>\% dplyr::mutate(model="7:2"),
  tmp3 \%>\% dplyr::mutate(model="14:1"),
) \%>\% dplyr::group_by(model,class)

if (interactive()) {
  plot_proportion(
    comp,
    date_labels="\%b \%y",
    mapping=ggplot2::aes(colour=model),
    raw=data
  )+ggplot2::facet_wrap(~class)
}

}
\concept{models}
