% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standalone-distributions.R
\name{rdiscgamma}
\alias{rdiscgamma}
\title{Random count data from a discrete gamma distribution}
\usage{
rdiscgamma(n, mean, sd, kappa)
}
\arguments{
\item{n}{number of observations}

\item{mean}{the mean value on the true scale (vectorised)}

\item{sd}{the standard deviation on the true scale (vectorised)}

\item{kappa}{a coefficient of variation. where 0 is no variability and 1 is
maximally variability (vectorised)}
}
\value{
an integer valued vector from a gamma distribution.
}
\description{
For count data that is under-dispersed we can use a gamma distribution rounded
to the nearest whole number. This is the same as the method of discretisation
in \code{make_gamma_ip}, and suits delay distributions where there is less variability
than can be represented with a Poisson or negative binomial distribution.
}
\examples{
rdiscgamma(10, 2, 1)
}
\seealso{
\code{\link[stats:GammaDist]{stats::rgamma()}}
}
\concept{distributions}
