% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reband-discrete-data.R
\name{reband_discrete}
\alias{reband_discrete}
\title{Reband any discrete distribution}
\usage{
reband_discrete(
  x,
  y,
  xout,
  xlim = c(0, NA),
  ytotal = c(0, sum(y)),
  digits = 0,
  labelling = c("positive_integer", "inclusive", "exclusive"),
  sep = "-"
)
}
\arguments{
\item{x}{a set of upper limits of bands, e.g. for age: 0-14;15-64;65-79;80+ is 15,65,80,NA}

\item{y}{a set of quantities for each band e.g. population figures}

\item{xout}{a set of new upper limits}

\item{xlim}{Upper and lower limits for x. if the last band is e.g 80+ in the input and we want to know the 85+
band in the output some kind of maximum upper limit is needed to interpolate
to.}

\item{ytotal}{upper and lower limits for y. If the interpolation values fall outside of
x then the minimum and maximum limits of y are given by this. This would be \code{c(0,1)}
for a probability distribution, for example.}

\item{digits}{if the \code{xout} value is continuous then how many significant figures
to put in the labels}

\item{labelling}{are the \code{xout} values interpretable as an \code{inclusive} upper limit, or
an \code{exclusive} upper limit, or as an upper limit of an `positive_integer`` quantity}

\item{sep}{separator for names e.g. \code{18-24} or \verb{18 to 24}}
}
\value{
a re-banded set of discrete values, guaranteed to sum to the same as \code{y}
}
\description{
e.g. age banded population, or a discrete probability distribution e.g.
a serial interval distribution. This method fits a monotonically increasing
spline to the cumulative distribution (including the upper and lower limits)
and interpolating using that spline to the new cut points.
}
\examples{

england_demographics = ukc19::uk_population_2019_by_10yr_age \%>\%
  dplyr::filter(name=="England")

ul = stringr::str_extract(england_demographics$class, "_([0-9]+)",group = 1) \%>\%
  as.numeric()
# ul is currently inclusive so add 1:
ul = ul + 1

tmp = reband_discrete(
  ul, england_demographics$population,
  c(5,10,15,40,80), xlim=c(0,120))

tmp

sum(tmp)
sum(england_demographics$population)
}
\concept{wrangling}
