% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale-time.R
\name{rescale_model}
\alias{rescale_model}
\title{Rescale a timeseries in the temporal dimension}
\usage{
rescale_model(df = i_timeseries, time_unit)
}
\arguments{
\item{df}{A data frame containing modelled output. This will modify the
following columns if present:

A dataframe containing the following columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item incidence.fit (double) - an estimate of the incidence rate on a log scale
\item incidence.se.fit (positive_double) - the standard error of the incidence rate estimate on a log scale
\item incidence.0.025 (positive_double) - lower confidence limit of the incidence rate (true scale)
\item incidence.0.5 (positive_double) - median estimate of the incidence rate (true scale)
\item incidence.0.975 (positive_double) - upper confidence limit of the incidence rate (true scale)
}

Any grouping allowed.

A dataframe containing the following columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item growth.fit (double) - an estimate of the growth rate
\item growth.se.fit (positive_double) - the standard error the growth rate
\item growth.0.025 (double) - lower confidence limit of the growth rate
\item growth.0.5 (double) - median estimate of the growth rate
\item growth.0.975 (double) - upper confidence limit of the growth rate
}

Any grouping allowed.

A dataframe containing the following columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item proportion.fit (double) - an estimate of the proportion on a logit scale
\item proportion.se.fit (positive_double) - the standard error of proportion estimate on a logit scale
\item proportion.0.025 (proportion) - lower confidence limit of proportion (true scale)
\item proportion.0.5 (proportion) - median estimate of proportion (true scale)
\item proportion.0.975 (proportion) - upper confidence limit of proportion (true scale)
\item relative.growth.fit (double) - an estimate of the relative growth rate
\item relative.growth.se.fit (positive_double) - the standard error the relative growth rate
\item relative.growth.0.025 (double) - lower confidence limit of the relative growth rate
\item relative.growth.0.5 (double) - median estimate of the relative growth rate
\item relative.growth.0.975 (double) - upper confidence limit of the relative growth rate
}

Any grouping allowed.}

\item{time_unit}{a \code{lubridate} period string such as "1 day"}
}
\value{
the same time series with different time unit, and adjusted incidence
and growth rate figures.
}
\description{
Sometimes we may have, for example, modelled incidence or growth rates on
weekly data resulting in cases per week and growth rate per week. We may wish
to use this to estimate the reproduction number, using algorithms that assume
a daily incidence. Not everything has a dependence on time, and things such
as proportions, or prevalence will not change.
}
\examples{

sim = sim_poisson_model(time_unit = "1 week")
incidence = sim \%>\% poisson_locfit_model(frequency = "1 day", deg = 2, window=5)
incidence2 = incidence \%>\% rescale_model(time_unit = "1 day")
incidence2 \%>\% dplyr::glimpse()

}
