% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-ggplot.R
\name{scale_x_logit}
\alias{scale_x_logit}
\title{A logit x scale}
\usage{
scale_x_logit(...)
}
\arguments{
\item{...}{Other arguments passed on to \verb{scale_(x|y)_continuous()}}
}
\value{
a ggplot scale
}
\description{
A logit x scale
}
\concept{vis}
