% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-estimates.R
\name{score_estimate}
\alias{score_estimate}
\title{Calculate scoring statistics from predictions using \code{scoringutils}.}
\usage{
score_estimate(est, obs, lags = NULL)
}
\arguments{
\item{est}{a dataframe of estimates of incidence, growth rate of reproduction
number based off a simulation or data with known parameters.}

\item{obs}{a dataframe of the ground truth, sharing the same grouping as
\code{est} with at least one column(s) named \code{XXX.obs} with \code{XXX} being one of
\code{rt},\code{growth} or \code{incidence} or any other column group predicted in \code{est}.}

\item{lags}{a data frame of estimate types and lags as output by \code{\link[=quantify_lag]{quantify_lag()}}
if multiple models are included then the columns must match those in \code{obs}.}
}
\value{
a dataframe of scoring metrics
}
\description{
This performs a range of quantile based, and if cumulative distribution
functions are available, continuous scoring metrics for each estimate time-point.
}
\examples{
tmp2 = test_bpm \%>\% sim_summarise_linelist()

withr::with_options(list("ggoutbreak.keep_cdf"=TRUE),{
   est = tmp2 \%>\% poisson_locfit_model() \%>\% rt_from_incidence()
})

obs = tmp2 \%>\% dplyr::mutate(rt.obs = rt.weighted)
score_estimate(est,obs) \%>\% dplyr::glimpse()

}
