% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic-times.R
\name{set_defaults}
\alias{set_defaults}
\alias{with_defaults}
\alias{set_default_start}
\alias{set_default_unit}
\title{Set or reset the default origin and unit for time periods}
\usage{
set_defaults(start_date, unit)

with_defaults(start_date, unit, expr)

set_default_start(date)

set_default_unit(unit)
}
\arguments{
\item{start_date}{the zero time date as something that can be coerced to a
date. If the \code{x} input is already a \code{time_period} and this is different to its
\code{start_date} then \code{x} will be recalibrated to use the new start date.}

\item{unit}{the length of one unit of time. This will be either a integer
number of days, or a specification such as "1 week", or another \code{time_period}.
If \code{x} is a \code{time_period}, and the unit is different to that of \code{x}
this will return a rescaled \code{time_period} using the new units.}

\item{expr}{an expression to evaluate with the defaults set to the provided
values.}
}
\value{
depending on the methods the original default start date / the
original default unit, a list of both or the result of evaluating the
expression \code{expr}
}
\description{
This function is generally not needed, and will be called automatically
when the first date conversion is performed. If no other information is
given then the default origin will be decided by the start of the first use of the
\code{time_period} class in a session. This helps to keep defaults consistent
during a single run, if they are not specified.
}
\section{Functions}{
\itemize{
\item \code{with_defaults()}: Set defaults temporarily and execute expression

\item \code{set_default_unit()}: Set the default unit only

}}
\examples{
# set default origin and cadence:
old = set_defaults("2025-01-01", "1 week")

# this sets the default for interpreting underqualified time_periods:
print(as.time_period(1:10))

# The default can always be overridden on a case by case basis:
print(as.time_period(1:10, unit="1 day"))

# or for a whole expression:
with_defaults("2020-01-01", "1 day", {
  print(as.time_period(1:10))
})

# components can be changed individually, firstly origin:
set_default_start("2025-01-01")
print(as.time_period(1:10))

# now cadence:
set_default_unit("1 day")
print(as.time_period(1:10))

# clear the values:
set_defaults(NULL,NULL)

# A sufficiently qualified call will set the defaults:
defined = as.time_period(as.Date("2024-01-01")+0:10*7, anchor="start")
inherit = as.time_period(0:10)

all(defined == inherit)

# restoring the original values (which might be null)
set_defaults(old)

}
\concept{time_period}
