% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-utils.R
\name{sim_events}
\alias{sim_events}
\title{Extract the events dataframe from a simulation output}
\usage{
sim_events(df)
}
\arguments{
\item{df}{the output of a \code{ggoutbreak} simulation}
}
\value{
an events dataframe
}
\description{
All the simulations should include details of major changes in the simulation
input parameters, particularly for step functions. This set of events can be
directly represented on a \code{ggoutbreak} plot using the \code{events} parameter
}
\examples{
sim_events(test_poisson_rt)
}
\concept{test}
