% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-utils.R
\name{sim_multinomial}
\alias{sim_multinomial}
\title{Generate a multinomial outbreak defined by per class growth rates and a
poisson model}
\usage{
sim_multinomial(
  changes = dplyr::tibble(t = c(0, 20, 40, 60, 80), variant1 = c(0.1, 0, -0.1, 0, 0.1),
    variant2 = c(0.15, 0.05, -0.05, -0.01, 0.05), variant3 = c(0, 0.05, -0.05, +0.05,
    -0.05), ),
  initial = c(100, 100, 100),
  time_unit = "1 day",
  ...
)
}
\arguments{
\item{changes}{a list of time points in column \code{t} and growth rates per week
per class, in other columns.}

\item{initial}{the size of the initial outbreak per class. There should be
one entry per class}

\item{time_unit}{e.g. a daily or weekly time series: "1 day", "7 days"}

\item{...}{
  Named arguments passed on to \code{\link[=sim_poisson_model]{sim_poisson_model}}\describe{
    \item{\code{fn_imports}}{a function that takes input vector \code{t} and returns the
number of imported cases at times \code{t}.}
\item{\code{seed}}{a random seed}
\item{\code{kappa}}{a dispersion parameter. 1 is no dispersion (compared to poisson), smaller values mean more dispersion.}
\item{\code{max_time}}{the desired length of the time series,}
\item{\code{fn_growth}}{a function that takes input vector \code{t} and returns the growth rates at times \code{t}}
}}
}
\value{
a case count time series including \code{class}, \code{count} and \code{time} columns
}
\description{
Generate a multinomial outbreak defined by per class growth rates and a
poisson model
}
\examples{
if (interactive()) {
  plot_counts(
    sim_multinomial() \%>\% dplyr::glimpse()
  )
}
}
\concept{test}
