% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-utils.R
\name{sim_poisson_Rt_model}
\alias{sim_poisson_Rt_model}
\title{Generate an outbreak case count series defined by Reproduction number using a poisson model.}
\usage{
sim_poisson_Rt_model(
  changes = dplyr::tibble(t = c(0, 40), rt = c(2.5, 0.8)),
  kappa = 1,
  max_time = 80,
  seed = Sys.time(),
  fn_Rt = cfg_step_fn(changes),
  fn_imports = ~ifelse(.x == 0, 30, 0),
  fn_ip = ~example_ip(),
  time_unit = "1 day"
)
}
\arguments{
\item{changes}{a dataframe holding change time points (\code{t}) and
reproduction number (\code{rt}) columns}

\item{kappa}{a dispersion parameter. 1 is no dispersion (compared to poisson), smaller values mean more dispersion.}

\item{max_time}{the desired length of the time series,}

\item{seed}{a random seed}

\item{fn_Rt}{a function that takes input vector \code{t} and returns the instantaneous reproduction
number at time \code{t}}

\item{fn_imports}{a function that takes input vector \code{t} and returns the
number of imported cases at times \code{t}.}

\item{fn_ip}{a function that takes input vector \code{t} and returns an
infectivity profile at times \code{t}.}

\item{time_unit}{e.g. a daily or weekly time series: "1 day", "7 days"}
}
\value{
A dataframe of case counts

A dataframe containing the following columns: 
\itemize{
\item statistic (character) - An identifier for the statistic, whether that be infections, admissions, deaths
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
}

Minimally grouped by: statistic (and other groupings may be present).
}
\description{
Generate an outbreak case count series defined by Reproduction number using a poisson model.
}
\examples{

tmp = sim_poisson_Rt_model(kappa=1, seed=100, fn_imports = ~ ifelse(.x \%in\% c(0,50),100,0))

if (interactive()) {
  ggplot2::ggplot(tmp,ggplot2::aes(x=time,y=count))+ggplot2::geom_point()+
    ggplot2::geom_line()
}

}
\concept{test}
