% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-utils.R
\name{sim_seir_model}
\alias{sim_seir_model}
\title{SEIR model with time-varying transmission parameter}
\usage{
sim_seir_model(
  changes = dplyr::tibble(t = c(0, 30), dBeta = c(1, 0.5)),
  mean_latent_period,
  mean_gen_time,
  R0 = 2.5,
  fn_dBeta = cfg_step_fn(changes),
  N = 10000,
  imports = 10,
  max_time = 104,
  seed = Sys.time()
)
}
\arguments{
\item{changes}{a dataframe holding change time points (\code{t}) and
the proportional increase in the transmission rate in the \code{dBeta} column}

\item{mean_latent_period}{Mean time from infection to becoming infectious (E to I),
assumed to be exponentially distributed.}

\item{mean_gen_time}{The average generation time (will be latent period+infectious duration).}

\item{R0}{the initial reproduction number}

\item{fn_dBeta}{A function of time \code{t} that returns the multiple of
transmission rate at that time point. Transmission (beta) at time zero is
defined by R0 and the infectious duration, This is multiplied at each time
by this factor.}

\item{N}{Total population size. Defaults to 10,000.}

\item{imports}{Initial number of infectious individuals. Defaults to 10.}

\item{max_time}{Maximum simulation time (in days or time units). Defaults to 100.}

\item{seed}{a random seed}
}
\value{
A dataframe containing the following columns:
\itemize{
\item statistic (character) - An identifier for the statistic, whether that be infections, admissions, deaths
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
}

Minimally grouped by: statistic (and other groupings allowed).
}
\description{
This function simulates an SEIR (Susceptible-Exposed-Infectious-Recovered) model
where the transmission rate (\code{beta}) varies over time.
}
\details{
The latent period (time from infection to becoming infectious) is assumed to be
exponentially distributed. The infectious period is derived from the given
mean of the generation time and the latent period.

The model assumes:
\itemize{
\item Latent period ~ Exponential(sigma), where sigma = 1 / \code{mean_latent_period}
\item Infectious period ~ Exponential(gamma), where gamma = 1 / (mean_gen_time - mean_latent_period)
\item Generation time ~ Exponential(sigma) + Exponential(gamma)
\item beta0 = R0 * gamma
\item Transmission rate beta(t) = fn_dBeta(t) * beta0
}

The generation time is defined as the sum of the latent and infectious periods.
}
\examples{
# Example: Lockdown after day 30

seir_output <- sim_seir_model(
  mean_latent_period = 4,
  mean_gen_time = 7,
  R0 = 2.5,
  fn_dBeta = function(t) ifelse(t < 30, 1, 0.5)
)

if (interactive()) {
  plot_ip(attr(seir_output,"ip"))
  plot_counts(seir_output)
}

}
\concept{test}
