% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation-utils.R
\name{sim_test_data}
\alias{sim_test_data}
\title{Generate a simple time-series of cases based on a growth rate step function}
\usage{
sim_test_data(ip = test_ip, duration = 500, period = 50)
}
\arguments{
\item{ip}{an infectivity profile. any uncertainty will be collapsed into the
central distribution.}

\item{duration}{the total length of the time-series}

\item{period}{the duration of each positive or negative growth phase}
}
\value{
a time series with \code{count}, \code{incidence}, \code{growth}, \code{rt},
\code{proportion} and \code{relative.growth} columns
}
\description{
This time-series has no statistical noise and is useful for testing things.
It has a fixed known value for infections and growth rate (fixed at 0.05 and
-0.05 per day), and a instantaneous reproduction number which is based on the
provided infectivity profile.  A fixed denominator gives a known proportion
and a relative growth rate that is the same as the growth rate.
}
\examples{
sim_test_data() \%>\% dplyr::glimpse()
}
\concept{test}
