% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_bpm}
\alias{test_bpm}
\title{An example of the linelist output of the branching process model simulation}
\format{
A dataframe containing the following columns:
\itemize{
\item time (as.time_period) - the time column
\item id (integer) - an id per individual
\item generation_interval (numeric) - the generation_interval column
\item infector (integer) - the infector id
\item generation (numeric) - the generation column
\item symptom_onset (logical) - the flag for onset of symptoms
\item symptom_onset_delay (numeric) - the time to onset of symptoms from infection
\item symptom_onset_time (as.time_period) - the time of symptom onset
}

333126 rows and 8 columns
}
\usage{
data(test_bpm)
}
\description{
This is generated using the \code{test_ip} infectivity profile and
also includes a delay to symptom onset which is a random gamma distributed
quantity with mean of 6 and sd of 2
}
\keyword{datasets}
