% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_delayed_observation}
\alias{test_delayed_observation}
\title{The delayed observation dataset}
\format{
A dataframe containing the following columns:
\itemize{
\item statistic (character) - the statistic column (only admissions)
\item obs_time (as.time_period) - the time of observation of the time-series
\item time (as.time_period) - the time of the data point in the time-series
\item count (integer) - the count of admissions
}

Grouped by: obs_time + statistic.

3321 rows and 4 columns
}
\usage{
data(test_delayed_observation)
}
\description{
This simulates what might be observed in an outbreak if there was on average
a 5 day delay on the reporting of hospital admissions.
The configuration of the outbreak is the same as \code{ggoutbreak::test_bpm}, but
this is summary data that describes the whole history of admissions that were
observed, when observed at any given time point. This is a triangular set
of data where the counts are right censored by the observation time.
}
\concept{datasets}
\keyword{datasets}
