% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_ip}
\alias{test_ip}
\title{A test infectivity profile generated from a set of discretised gamma
distributions with parameters mean 5 (95\% CI 4-6) and sd 2 (95\% CI 1.5-2.5).}
\format{
A dataframe containing the following columns:
\itemize{
\item boot (anything + default(1)) - a bootstrap identifier
\item probability (proportion) - the probability of infection between previous time period until \code{time}
\item tau (numeric) - the time index this probability relates to (in days)
\item a0 (numeric) - the beginning of the time period
\item a1 (numeric) - the end of the time period
}

Grouped by: boot.

2000 rows and 5 columns
}
\usage{
data(test_ip)
}
\description{
A test infectivity profile generated from a set of discretised gamma
distributions with parameters mean 5 (95\% CI 4-6) and sd 2 (95\% CI 1.5-2.5).
}
\concept{datasets}
\keyword{datasets}
