% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_poisson_growth_rate}
\alias{test_poisson_growth_rate}
\title{A simulation dataset determined by a step function of growth rates. This is
useful for demonstrating growth rate estimators.}
\format{
A dataframe containing the following columns:
\itemize{
\item time (as.time_period) - the time column
\item growth (numeric) - the time varying growth rate column (input parameter)
\item imports (numeric) - the imports column
\item rate (numeric) - the poisson rate column
\item count (integer) - the sampled count column
\item statistic (character) - the statistic column (infections)
}

Minimally grouped by: statistic (and other groupings allowed).

105 rows and 6 columns
}
\usage{
data(test_poisson_growth_rate)
}
\description{
A simulation dataset determined by a step function of growth rates. This is
useful for demonstrating growth rate estimators.
}
\concept{datasets}
\keyword{datasets}
