% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_poisson_rt}
\alias{test_poisson_rt}
\title{An example of the linelist output of the poisson model simulation with defined
$R_t$}
\format{
A dataframe containing the following columns:
\itemize{
\item time (as.time_period) - the time column
\item rt (numeric) - the time varying rt column (parameters)
\item imports (numeric) - the imports column
\item rate (numeric) - the poisson rate column (underlying infection rate)
\item count (integer) - the count column
\item statistic (character) - the statistic column
}

81 rows and 6 columns
}
\usage{
data(test_poisson_rt)
}
\description{
This is generated using the \code{test_ip} infectivity profile
}
\concept{datasets}
