% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_poisson_rt_2class}
\alias{test_poisson_rt_2class}
\title{The test_poisson_rt_2class dataset}
\format{
A dataframe containing the following columns:
\itemize{
\item time (as.time_period) - the time column
\item rt (numeric) - the rt column
\item imports (numeric) - the imports column
\item rate (numeric) - the rate column
\item count (integer) - the count column
\item statistic (character) - the statistic column
\item class (enum(\code{one},\code{two})) - the class column
\item denom (integer) - the denom column
}

Minimally grouped by: class (and other groupings allowed).

322 rows and 8 columns
}
\usage{
data(test_poisson_rt_2class)
}
\description{
The test_poisson_rt_2class dataset
}
\concept{datasets}
\keyword{datasets}
