% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_poisson_rt_smooth}
\alias{test_poisson_rt_smooth}
\title{Output of a poisson model simulation with a
smooth function for $R_t$ defined as \verb{R(t) = e^(sin(t/80*pi)^4-0.25))}. This
is a relatively unchallenging test data set that should not pose a problem
for smooth estimators.}
\format{
A dataframe containing the following columns:
\itemize{
\item time (as.time_period) - the time column
\item rt (numeric) - the time varying rt column (parameters)
\item imports (numeric) - the imports column
\item rate (numeric) - the poisson rate column (underlying infection rate)
\item count (integer) - the count column
\item statistic (character) - the statistic column
}

Minimally grouped by: statistic (and other groupings allowed).

161 rows and 6 columns
}
\usage{
data(test_poisson_rt_smooth)
}
\description{
This is generated using the central value of the \code{test_ip} infectivity profile
}
\concept{datasets}
\keyword{datasets}
