% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_serial}
\alias{test_serial}
\title{A serial interval estimated from simulated data}
\format{
A dataframe containing the following columns:
\itemize{
\item tau (numeric) - the time delay between symptoms in infector and infectee
\item a0 (numeric) - the a0 column
\item a1 (numeric) - the a1 column
\item probability (numeric) - the probability column
\item boot (integer) - the boot column
}

Minimally grouped by: boot (and other groupings allowed).

2166 rows and 5 columns
}
\usage{
data(test_serial)
}
\description{
This serial interval is resampled from the first 1000 patients in the
\code{test_bpm} dataset for whom both infector and infectee has symptoms. These
patients are generated with a symptom delay of mean 6 days and SD 2 from
infection (discrete under-dispersed gamma) and an infectivity profile with
mean 5 days and SD 2 as defined in \code{test_ip} dataset. This serial interval
if relevant to the estimation of $R_t$ from symptomatic case counts but
includes negative times, and cannot be used with \code{EpiEstim}.
}
\concept{datasets}
