% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_ts}
\alias{test_ts}
\title{The test_ts dataset}
\format{
A dataframe containing the following columns:
\itemize{
\item time (as.time_period) - the time column
\item growth (numeric) - the growth column
\item incidence (numeric) - the incidence column
\item rt (numeric) - the rt column
\item denom (numeric) - the denom column
\item proportion (numeric) - the proportion column
\item relative.growth (numeric) - the relative.growth column
\item count (numeric) - the count column
}

Any grouping allowed.

501 rows and 8 columns
}
\usage{
data(test_ts)
}
\description{
The test_ts dataset
}
\keyword{datasets}
