% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time-summarise.R
\name{time_aggregate}
\alias{time_aggregate}
\title{Aggregate time series data preserving the time series}
\usage{
time_aggregate(
  df = i_timestamped,
  ...,
  .groups = NULL,
  .cols = NULL,
  .fns = NULL
)
}
\arguments{
\item{df}{an optionally grouped time series. Grouping should not include the time
column. The grouping works differently from \code{dplyr::summarise} in that the last level
of non-time groups is lost in this operation, so the subgroup you wish to aggregate
should be included in the grouping.}

\item{...}{A set of \code{dplyr::summarise} statements, or additional parameters
for \code{.fns}}

\item{.groups}{as per \code{dplyr::summarise}}

\item{.cols}{Optional dplyr column specification for the data that will be
summarised. This is useful if there are lots of columns you want to
summarise using the same function (e.g. \code{sum} usually). \code{.cols} is passed to
a \code{dplyr::across} call. if
\code{.fns} is given and the \code{.cols} parameter is not specified then the columns
to summarise are automatically identified. In doing this any \code{Date} columns
are dropped. If this in not what you want then \code{.cols} or \code{...} must be given}

\item{.fns}{Optional a set of function specifications as per \code{dplyr::across}}
}
\value{
the summarised time series preserving the \code{time} column, and with the grouping
structure involving one fewer levels than the input
}
\description{
This function operates on timeseries data not linelists (see \code{\link[=time_summarise]{time_summarise()}})
for line lists. If a very granular timeseries is regrouped and this function
is applied the resulting dataframe will be
}
\examples{
example_england_covid_by_age() \%>\%
  time_aggregate(count = sum(count), denom = sum(denom)) \%>\%
  dplyr::glimpse()

example_england_covid_by_age() \%>\%
  time_aggregate(.fns=mean, .cols=dplyr::where(is.numeric)) \%>\%
  dplyr::glimpse()
}
\concept{time_period}
