% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-import.R
\name{timeseries}
\alias{timeseries}
\alias{timeseries.default}
\alias{timeseries.incidence2}
\alias{timeseries.data.frame}
\alias{timeseries.incidence}
\title{Coerce an object to a \code{ggoutbreak} compatible time series dataframe}
\usage{
timeseries(x, ...)

\method{timeseries}{default}(x, ...)

\method{timeseries}{incidence2}(x, ...)

\method{timeseries}{data.frame}(
  x,
  ...,
  date = NULL,
  count = NULL,
  denom = NULL,
  population = NULL,
  class = NULL,
  declutter = FALSE
)

\method{timeseries}{incidence}(x, ..., declutter = FALSE)
}
\arguments{
\item{x}{An object to coerce e.g. a data frame}

\item{...}{
  Named arguments passed on to \code{\link[=as.time_period]{as.time_period}}\describe{
    \item{\code{unit}}{the length of one unit of time. This will be either a integer
number of days, or a specification such as "1 week", or another \code{time_period}.
If \code{x} is a \code{time_period}, and the unit is different to that of \code{x}
this will return a rescaled \code{time_period} using the new units.}
\item{\code{start_date}}{the zero time date as something that can be coerced to a
date. If the \code{x} input is already a \code{time_period} and this is different to its
\code{start_date} then \code{x} will be recalibrated to use the new start date.}
\item{\code{anchor}}{only relevant if \code{x} is a vector of dates, this is a date, or
\code{"start"} or \code{"end"} or a weekday name e.g.
\code{"mon"}. With the vector of dates in \code{x} it will use this anchor to find a reference date for
the time-series. If not provided then the current defaults will be used.
(see \code{\link[=set_defaults]{set_defaults()}})}
}}

\item{date}{R expression defining a date column (as a \code{Date} - optional)}

\item{count}{R expression defining a count column (as an \code{int} - optional)}

\item{denom}{R expression defining a denominator column (as an \code{int} - optional)}

\item{population}{R expression defining a population column (as an \code{numeric} - optional)}

\item{class}{R expression defining a class column (as an \code{factor} - optional)}

\item{declutter}{logical flag. If \code{TRUE} then unnecessary original columns will be dropped}
}
\value{
minimally a case count dataframe A dataframe containing the following columns:
\itemize{
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
}

Any grouping allowed.
}
\description{
Coerce an object to a \code{ggoutbreak} compatible time series dataframe
}
\examples{

data("mers_2014_15", package="EpiEstim")

# Use an expression to generate a case count:
# N.B. complex column names need to be surrounded with bcakticks like this:

tmp = mers_2014_15$incidence \%>\%
  timeseries(
    date = `mers$dates`,
    count = local+imported
  ) \%>\%
  dplyr::glimpse()

if (interactive()) {
  ip = make_fixed_ip(mean = mers_2014_15$si$mean_si, sd = mers_2014_15$si$std_si)
  tmp \%>\% poisson_locfit_model(ip=ip,window=14) \%>\% plot_rt()
}


# remove columns not needed by `ggoutbreak`
outbreaks::dengue_fais_2011 \%>\%
  timeseries(date = onset_date, count = value, declutter = TRUE) \%>\%
  dplyr::glimpse()


# date column already exists,  could use `onset` or `death` as count
outbreaks::ebola_kikwit_1995 \%>\%
  timeseries(count = onset) \%>\%
  dplyr::glimpse()

# This data set needs grouping to make it a unique timeseries:
ukc19::ltla_cases \%>\%
  timeseries(anchor="start", unit="1 day") \%>\%
  dplyr::glimpse()

# from an incidence object:
if (requireNamespace("incidence",quietly = TRUE)) {

  onset = outbreaks::ebola_sim$linelist$date_of_onset
  sex = outbreaks::ebola_sim$linelist$gender
  inc.week.gender = incidence::incidence(onset, interval = 7, groups = sex, standard = FALSE)

  inc.week.gender \%>\% timeseries() \%>\% dplyr::glimpse()

  d = Sys.Date() + sample(-3:10, 10, replace = TRUE)
  di = incidence::incidence(d, interval = "week", first_date = Sys.Date() - 10, standard = TRUE)
  di \%>\% timeseries(declutter=TRUE) \%>\% dplyr::glimpse()
}

if (requireNamespace("incidence2",quietly = TRUE)) {
  dat = outbreaks::ebola_sim_clean$linelist
  x = incidence2::incidence(dat, "date_of_onset", groups = c("gender", "hospital"), interval="week")
  x \%>\% timeseries() \%>\% dplyr::glimpse()
}

}
