% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periodic-times.R
\name{type.time_period}
\alias{type.time_period}
\title{Type coercion to a \code{time_period} class}
\usage{
type.time_period(x)
}
\arguments{
\item{x}{a vector to be coerced to a time period}
}
\value{
a \code{time_period} or an error
}
\description{
Type coercion to a \code{time_period} class
}
\examples{
type.time_period(1:100)
}
