% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-rt-wallinga.R
\name{wallinga_lipsitch}
\alias{wallinga_lipsitch}
\title{Calculate the reproduction number from a growth rate estimate and an infectivity profile}
\usage{
wallinga_lipsitch(
  r,
  y = i_empirical_ip,
  a1 = seq(0.5, length.out = length(y)),
  a0 = dplyr::lag(a1, default = 0)
)
}
\arguments{
\item{r}{a growth rate (may be a vector)}

\item{y}{an empirical infectivity profile either as a probability vector or as a dataframe of format:
A dataframe containing the following columns:
\itemize{
\item boot (anything + default(1)) - a bootstrap identifier
\item probability (proportion) - the probability of new event during this period.
\item a0 (double) - the beginning of the time period (in days)
\item a1 (double) - the end of the time period (in days)
}

Minimally grouped by: boot (and other groupings allowed).
A default value is defined.}

\item{a1}{the end time of the infectivity profile probability estimate (defaults to 0.5,1.5,2.5,...).}

\item{a0}{the start time of the infectivity profile probability estimate (defaults to 0,0.5,1.5,...).}
}
\value{
a reproduction number estimate based on \code{r}
}
\description{
This function uses a single empirical distribution for the infectivity
profile aka generation time
}
\examples{
wallinga_lipsitch(r=seq(-0.1,0.1,length.out=9), y=stats::dgamma(1:50, 5,2))
}
\concept{models}
