% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-gam.R
\name{gam_delayed_reporting}
\alias{gam_delayed_reporting}
\title{Delayed GAM reporting model function generator}
\usage{
gam_delayed_reporting(
  window,
  max_delay = 40,
  ...,
  knots_fn = ~gam_knots(.x, window, ...)
)
}
\arguments{
\item{window}{controls the knot spacing in the GAM (if the default)}

\item{max_delay}{the maximum delay we expect to model}

\item{...}{
  Named arguments passed on to \code{\link[=gam_knots]{gam_knots}}\describe{
    \item{\code{data}}{the function will be called with incidence data - a dataframe with columns:
\itemize{
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
}

Any grouping allowed.
}
\item{\code{k}}{alternative to \code{window}, if \code{k} is given then the behaviour of the
knots will be similar to the default \code{mgcv::s(...,k=...)} parameter.}
\item{\code{...}}{currently not used}
}}

\item{knots_fn}{a function that takes the data as an input and returns
a set of integers as time points for GAM knots, for \code{s(time)} term. The
default here provides a roughly equally spaced grid determined by \code{window},
by a user supplied function could do anything. The input this function is
the raw dataframe of data that will be considered for one model fit. It is
guaranteed to have at least a \code{time} and \code{count} column. It is possible to}
}
\value{
a list with 2 entries - \code{model_fn} and \code{predict} suitable as the
input for \code{poisson_gam_model(model_fn = ..., predict=...)}.
}
\description{
This function is used to configure a delayed reporting GAM model. The model
is of the form:
}
\details{
\code{count ~ s(time, bs = "cr", k = length(kts)) + s(log(tau), k = 4, pc = max_delay)}

where \code{tau} is the difference between time series observation time and the
time of the data point in the time series, and we have multiple observations
of the same time series. This function helps specify the knots of the GAM and
the maximum expected delay
}
\examples{

data = test_delayed_observation \%>\% dplyr::group_by(obs_time)
cfg = gam_delayed_reporting(14,40)
fit = cfg$model_fn(data)
summary(fit)

}
\concept{models}
