% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-raw-ts.R
\name{plot_cases}
\alias{plot_cases}
\title{Plot a line-list of cases as a histogram}
\usage{
plot_cases(
  raw,
  ...,
  mapping = .check_for_aes(raw, ..., class_aes = "fill"),
  events = i_events
)
}
\arguments{
\item{raw}{The raw case data either as a summarised count or as a line-list - EITHER: a dataframe with columns:
\itemize{
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
}

Any grouping allowed.

OR with columns:
\itemize{
\item time (ggoutbreak::time_period) - A set of events with a timestamp as a `time_period`
}

Any grouping allowed.

}

\item{...}{
  Named arguments passed on to \code{\link[=geom_events]{geom_events}}\describe{
    \item{\code{events}}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}}

\item{mapping}{a \code{ggplot2::aes} mapping. Most importantly setting the \code{fill}
to something if there are multiple types of event in the plot. If a \code{class}
column is present the mapping will default to using this.}

\item{events}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}
\value{
a ggplot object
}
\description{
Plot a line-list of cases as a histogram
}
\examples{

with_defaults("2025-01-01" ,"1 day", {
  tmp = dplyr::tibble(
    time = as.time_period( rexpgrowth(100,0.05,40)),
    class = rep(c("one","two","three"), length.out=100)
  )
})

if(interactive()) {
  plot_cases(tmp, mapping=ggplot2::aes(fill = class),linewidth=0.1,colour="white")
}
}
\concept{vis}
