% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-raw-ts.R
\name{plot_proportions_data}
\alias{plot_proportions_data}
\title{Plot a raw case count proportion timeseries}
\usage{
plot_proportions_data(
  raw = i_proportion_data,
  ...,
  mapping = .check_for_aes(raw, ...),
  events = i_events
)
}
\arguments{
\item{raw}{The raw count and denominator data - a dataframe with columns:
\itemize{
\item denom (positive_integer) - Total test counts associated with the specified time frame
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
}

Any grouping allowed.
}

\item{...}{
  Named arguments passed on to \code{\link[=geom_events]{geom_events}}\describe{
    \item{\code{events}}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}}

\item{mapping}{a \code{ggplot2::aes} mapping. Most importantly setting the \code{colour}
to something if there are multiple incidence timeseries in the plot}

\item{events}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}
\value{
a ggplot object
}
\description{
Plot a raw case count proportion timeseries
}
\examples{
# example code

tmp = dplyr::tibble(
  time = as.time_period(1:10, "1 day"),
  count = 101:110
) \%>\% dplyr::mutate(
  denom = count*time
)

if(interactive()) {
  plot_proportions_data(tmp)+ggplot2::geom_line()
}

}
\concept{vis}
