% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rt-ts.R
\name{plot_rt}
\alias{plot_rt}
\title{Reproduction number timeseries diagram}
\usage{
plot_rt(
  modelled = i_reproduction_number,
  ...,
  mapping = .check_for_aes(modelled, ...),
  events = i_events
)
}
\arguments{
\item{modelled}{the modelled Rt estimate - a dataframe with columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item rt.fit (double) - an estimate of the reproduction number
\item rt.se.fit (positive_double) - the standard error of the reproduction number
\item rt.0.025 (double) - lower confidence limit of the reproduction number
\item rt.0.5 (double) - median estimate of the reproduction number
\item rt.0.975 (double) - upper confidence limit of the reproduction number
}

Any grouping allowed.
}

\item{...}{
  Named arguments passed on to \code{\link[=geom_events]{geom_events}}\describe{
    \item{\code{events}}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}
  Named arguments passed on to \code{\link[=geom_truth]{geom_truth}}\describe{
    \item{\code{true_df}}{a data frame with a \code{time_period} column called time and a
value column (name given in \code{true_col}). This is optional and will be picked
up from the \code{raw} parameter if given.}
\item{\code{true_col}}{the column name / expression of the true value}
\item{\code{true_fmt}}{a list of ggplot formatting to apply to the true value timeseries}
\item{\code{...}}{
  Named arguments passed on to \code{\link[=as.time_period]{as.time_period}}\describe{
    \item{\code{unit}}{the length of one unit of time. This will be either a integer
number of days, or a specification such as "1 week", or another \code{time_period}.
If \code{x} is a \code{time_period}, and the unit is different to that of \code{x}
this will return a rescaled \code{time_period} using the new units.}
\item{\code{start_date}}{the zero time date as something that can be coerced to a
date. If the \code{x} input is already a \code{time_period} and this is different to its
\code{start_date} then \code{x} will be recalibrated to use the new start date.}
\item{\code{anchor}}{only relevant if \code{x} is a vector of dates, this is a date, or
\code{"start"} or \code{"end"} or a weekday name e.g.
\code{"mon"}. With the vector of dates in \code{x} it will use this anchor to find a reference date for
the time-series. If not provided then the current defaults will be used.
(see \code{\link[=set_defaults]{set_defaults()}})}
}}
}}

\item{mapping}{a \code{ggplot2::aes} mapping. Most importantly setting the \code{colour}
to something if there are multiple incidence time series in the plot}

\item{events}{Significant events or time spans - a dataframe with columns:
\itemize{
\item label (character) - the event label
\item start (date) - the start date, or the date of the event
\item end (date) - the end date or NA if a single event
}

Any grouping allowed.

A default value is defined.
}
}
\value{
a ggplot timeseries
}
\description{
Reproduction number timeseries diagram
}
\examples{
# example code
if (interactive()) {

tmp2 = example_poisson_locfit() \%>\%
  dplyr::filter(as.Date(time) >= "2021-01-01" & as.Date(time) < "2022-01-01") \%>\%
  rt_from_incidence(ip = example_ganyani_ip())

# comparing RT from growth rates with England consensus Rt
# (N.B. offset by 14 days to align with estimates):

plot_rt(tmp2,colour="blue")+
  ggplot2::geom_errorbar(
    data= ukc19::spim_consensus \%>\%
      dplyr::filter(date-14 >= "2021-01-01" & date-14 < "2022-01-01"),
    mapping=ggplot2::aes(x=date-14,ymin=rt.low,ymax=rt.high),
    colour="red")

}
}
\concept{vis}
