% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-rt-epiestim.R
\name{rt_epiestim}
\alias{rt_epiestim}
\title{\code{EpiEstim} reproduction number}
\usage{
rt_epiestim(
  df = i_incidence_input,
  ip = i_discrete_ip,
  bootstraps = 2000,
  window = 14,
  mean_prior = 1,
  std_prior = 2,
  ...
)
}
\arguments{
\item{df}{Count data. Extra groups are allowed.

A dataframe containing the following columns: 
\itemize{
\item count (positive_integer) - Positive case counts associated with the specified time frame
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
}

Ungrouped.}

\item{ip}{infectivity profile

A dataframe containing the following columns: 
\itemize{
\item boot (anything + default(1)) - a bootstrap identifier
\item probability (proportion) - the probability of new event during this period.
\item tau (integer + complete) - the days since the index event.
}

Must be grouped by: boot (exactly).

A default value is defined.}

\item{bootstraps}{\itemize{
\item the number of bootstraps to take to calculate for each point.
}}

\item{window}{\itemize{
\item the width of the \code{EpiEstim} window
}}

\item{mean_prior}{the prior for the $R_t$ estimate. When sample size is low the
$R_t$ estimate will revert to this prior. In \code{EpiEstim} the default is a high
number to allow detection of insufficient data but this tends to create
anomalies in the early part of infection time series. A possible value is $R_0$
but in fact this also will be a poor choice for the value of $R_t$ when case
numbers drop to a low value.}

\item{std_prior}{the prior for the $R_t$ SD.}

\item{...}{not used}
}
\value{
A dataframe containing the following columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item rt.fit (double) - an estimate of the reproduction number
\item rt.se.fit (positive_double) - the standard error of the reproduction number
\item rt.0.025 (double) - lower confidence limit of the reproduction number
\item rt.0.5 (double) - median estimate of the reproduction number
\item rt.0.975 (double) - upper confidence limit of the reproduction number
}

Any grouping allowed.
}
\description{
Calculate a reproduction number estimate from incidence data using the
\code{EpiEstim} library and an empirical generation time distribution. This uses
resampling to transmit uncertainty in generation time estimates. This is
quite slow for each time series depending on the number of bootstraps and
samples in the infectivity profile.
}
\details{
This will calculate a reproduction number for each group in the input dataframe.
}
\examples{
tmp = ggoutbreak::england_covid \%>\%
  time_aggregate(count=sum(count))

if (interactive()) {
  # not run due to long running
  tmp2 = tmp \%>\% rt_epiestim()
}


}
\concept{models}
