% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator-rt-incidence.R
\name{rt_from_incidence}
\alias{rt_from_incidence}
\title{Reproduction number from modelled incidence}
\usage{
rt_from_incidence(
  df = i_incidence_model,
  ip = i_discrete_ip,
  approx = FALSE,
  .progress = interactive()
)
}
\arguments{
\item{df}{modelled incidence estimate - a dataframe with columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a `time_period`
\item incidence.fit (double) - an estimate of the incidence rate on a log scale
\item incidence.se.fit (positive_double) - the standard error of the incidence rate estimate on a log scale
\item incidence.0.025 (positive_double) - lower confidence limit of the incidence rate (true scale)
\item incidence.0.5 (positive_double) - median estimate of the incidence rate (true scale)
\item incidence.0.975 (positive_double) - upper confidence limit of the incidence rate (true scale)
}

Any grouping allowed.
}

\item{ip}{an infectivity profile (aka generation time distribution) - a dataframe with columns:
\itemize{
\item boot (anything + default(1)) - a bootstrap identifier
\item probability (proportion) - the probability of new event during this period.
\item tau (integer + complete) - the days since the index event.
}

Minimally grouped by: boot (and other groupings allowed).

A default value is defined.
}

\item{approx}{use a faster, but approximate, estimate of quantiles}

\item{.progress}{show a CLI progress bar}
}
\value{
A dataframe containing the following columns:
\itemize{
\item time (ggoutbreak::time_period + group_unique) - A (usually complete) set of singular observations per unit time as a \code{time_period}
\item rt.fit (double) - an estimate of the reproduction number
\item rt.se.fit (positive_double) - the standard error of the reproduction number
\item rt.0.025 (double) - lower confidence limit of the reproduction number
\item rt.0.5 (double) - median estimate of the reproduction number
\item rt.0.975 (double) - upper confidence limit of the reproduction number
}

Any grouping allowed.
}
\description{
Calculate a reproduction number estimate from modelled incidence using the
methods described in the vignette "Estimating the reproduction number from
modelled incidence" and using a set of empirical generation time
distributions. This assumes that modelled incidence has the same time unit as
the \code{ip} distribution, and that this is daily, if this is not the case then
\code{\link[=rescale_model]{rescale_model()}} may be able to fix it.
}
\examples{
tmp = ggoutbreak::england_covid \%>\%
  dplyr::filter(date < "2021-01-01") \%>\%
  time_aggregate(count=sum(count)) \%>\%
  poisson_locfit_model() \%>\%
  rt_from_incidence()


if (interactive()) {
  plot_rt(tmp, date_labels="\%b \%y") \%above\%
   ggplot2::geom_errorbar(
     data=england_consensus_rt \%>\% dplyr::filter(date < "2021-01-01"),
     mapping=ggplot2::aes(x=date-14,ymin=low,ymax=high),colour="grey60")+
   ggplot2::coord_cartesian(ylim=c(0.5,1.75),xlim=as.Date(c("2020-05-01",NA)))
}

}
\concept{models}
