% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score-estimates.R
\name{score_estimate}
\alias{score_estimate}
\title{Calculate scoring statistics from predictions.}
\usage{
score_estimate(
  est,
  obs,
  lags = NULL,
  summarise_by = est \%>\% dplyr::groups(),
  bootstraps = 1000,
  raw_bootstraps = FALSE,
  seed = 100
)
}
\arguments{
\item{est}{a dataframe of estimates of incidence, growth rate of reproduction
number based off a simulation or data with known parameters. Each group in
\code{est} is expected to contain multiple estimates and each group is scored
separately. Estimates in \code{est} must be in the form of a column named
\code{XXX.cdf} containing a cumulative distribution function for the estimate
and \code{XXX.link} containing a link function specification (one if
\code{identity},\code{log} or \code{logit}). These are not generated by default by the
\code{ggoutbreak} estimators but are triggered by setting the option:
\code{options("ggoutbreak.keep_cdf"=TRUE)} before running the estimator. The
CDFs generated will be analytical, if the estimator generates a
parametrised output (or a mixture thereof), empirical if the estimator uses
resampling, or inferred if the estimator produces quantiles only.}

\item{obs}{a dataframe of the ground truth, sharing the same grouping and
columns as \code{est} with at least one column(s) named \code{XXX.obs} with \code{XXX}
being e.g. \code{rt},\code{growth} or \code{incidence} or any other column group
predicted in \code{est} (i.e. if \code{obs} has a column \code{XXX.obs}, \code{est} must have
one called \code{XXX.cdf}).}

\item{lags}{a data frame of estimate types and lags as output by
\code{\link[=quantify_lag]{quantify_lag()}} if multiple models are included then the columns must
match those in \code{obs}. It must have 2 columns, one called \code{estimate} with
values matching \code{incidence},\code{rt},\code{growth},\code{proportion},\code{relative.growth},
and a \code{lag} column, with (whole) number of days.}

\item{summarise_by}{by default every group is treated separately. This can be
overridden with a \code{dplyr} specification of the groupings we want to
see in the final summarised output (e.g. if we want to differentiate
performance on a particular type of scenario or timeframe). If this is
exactly \code{FALSE} the function will return all the raw point estimates.}

\item{bootstraps}{the number of bootstrap replicates to draw for assessing
metric confidence. If FALSE then no bootstrapping will be done and the
metrics returned will have no confidence intervals.}

\item{raw_bootstraps}{(defaults to FALSE) return the summary metrics for
each bootstrap rather than the quantiles of the summary metrics.}

\item{seed}{a random seed for reproducibility}
}
\value{
a dataframe of scoring metrics, with one row per group. This includes
the following columns:
\itemize{
\item \code{mean_quantile_bias} - the average of the universal residuals. Lower values
are better.
\item \code{mean_trans_bias} - the bias on the link function scale.
\item link - the link function
\item \code{mean_bias} - the bias on the natural scale (which may be interpreted as
additive or multiplicative depending on the link)
\item \code{pit_was} - an unadjusted probability integral transform histogram
Wasserstein distance from the uniform (lower values are better).
\item \code{unbiased_pit_was} - an PIT Wasserstein distance from the uniform, adjusted
for estimator bias (lower values are better). This is a measure of
calibration.
\item \code{directed_pit_was} - a PIT Wasserstein distance from the uniform, directed
away from the centre, adjusted for estimator bias (values closer to zero
are better, positive values indicate overconfidence, and negative values
excessively conservative estimates).
\item \code{percent_iqr_coverage} - the percentage of estimators that include the true
value in their IQR. For a perfectly calibrated estimate this should be 0.5.
Lower values reflect overconfidence, higher values reflect excessively
conservative estimates. This is a measure of calibration but is influenced
by bias.
\item \code{unbiased_percent_iqr_coverage} - the percentage of estimators that include
the true value in their IQR once adjusted for bias. This should be 0.5. This
is a measure of calibration, and tells you which direction (smaller numbers
are over-confident, larger values excessively conservative).
\item \code{mean_prediction_interval_width_50} - the prediction interval width is a
measure of sharpness (smaller values are sharper). Sharper estimators are
superior if they are unbiased and well calibrated.
\item \code{mean_crps} - the mean value of the continuous rank probability score for
each point estimate (lower values are better)
\item \code{mean_unbiased_crps} - the mean value of the continuous rank probability
score for each point estimate assessed after adjustment for bias (lower
values are better)
\item \code{threshold_misclassification_probability} - if a metric has a natural threshold
like 1 for Rt then this measures how probable it is that the estimate will
propose the epidemic is shrinking when it is growing and vice versa. Lower is
better
}

other outputs are possible if \code{summarise_by} is false.
}
\description{
This performs a range of continuous scoring metrics for each estimate
time-point using cumulative distribution functions for each estimate. Point
quality metrics are calculated for each estimate provided and summarised.
Summarisation is performed using bootstrap resampling to generate confidence
intervals for summary statistics, which are presented as median +/1 95\% CI.
}
\examples{
data = example_poisson_rt_smooth()

pipeline = ~ .x \%>\% poisson_locfit_model(ip = .y, quick=TRUE)
lags = quantify_lag(pipeline, ip = example_ip())

withr::with_options(list("ggoutbreak.keep_cdf"=TRUE),{
   est = data \%>\% poisson_locfit_model(ip = example_ip(), quick=TRUE)
})

if (interactive()) plot_rt(est)+sim_geom_function(data, colour="red")

obs = data \%>\% dplyr::mutate(rt.obs = rt, incidence.obs = rate)
score_estimate(est,obs,lags) \%>\% dplyr::glimpse()

}
\concept{test}
